anise::almanac

Module metaload

Structs

MetaAlmanac

A structure to set up an Almanac, with automatic downloading, local storage, checksum checking, and more.

MetaFile

MetaFile allows downloading a remote file from a URL (http, https only), and interpolation of paths in environment variable using the Dhall syntax env:MY_ENV_VAR.

Enums

MetaAlmanacError