anise::almanac

Struct Almanac

pub struct Almanac {
    pub spk_data: IndexMap<String, SPK>,
    pub bpc_data: IndexMap<String, BPC>,
    pub planetary_data: PlanetaryDataSet,
    pub spacecraft_data: SpacecraftDataSet,
    pub euler_param_data: EulerParameterDataSet,
    pub location_data: LocationDataSet,
}

An Almanac contains all of the loaded SPICE and ANISE data. It is the context for all computations.

:type path: str :rtype: Almanac

Fields
spk_data: IndexMap<String, SPK>

NAIF SPK is kept unchanged

bpc_data: IndexMap<String, BPC>

NAIF BPC is kept unchanged

planetary_data: PlanetaryDataSet

Dataset of planetary data

spacecraft_data: SpacecraftDataSet

Dataset of spacecraft data

euler_param_data: EulerParameterDataSet

Dataset of euler parameters

location_data: LocationDataSet

Dataset of locations

Implementations

impl Almanac

pub fn azimuth_elevation_range_sez( &self, rx: Orbit, tx: Orbit, obstructing_body: Option<Frame>, ab_corr: Option<Aberration>, ) -> AlmanacResult<AzElRange>

Computes the azimuth (in degrees), elevation (in degrees), and range (in kilometers) of the receiver state (rx) seen from the transmitter state (tx), once converted into the SEZ frame of the transmitter.

Warning

The obstructing body should be a tri-axial ellipsoid body, e.g. IAU_MOON_FRAME.

Algorithm

If any obstructing_bodies are provided, ensure that none of these are obstructing the line of sight between the receiver and transmitter.

Compute the SEZ (South East Zenith) frame of the transmitter.

Rotate the receiver position vector into the transmitter SEZ frame.

Rotate the transmitter position vector into that same SEZ frame.

Compute the range as the norm of the difference between these two position vectors.

Compute the elevation, and ensure it is between +/- 180 degrees.

Compute the azimuth with a quadrant check, and ensure it is between 0 and 360 degrees.

pub fn azimuth_elevation_range_sez_from_location_id( &self, rx: Orbit, location_id: i32, obstructing_body: Option<Frame>, ab_corr: Option<Aberration>, ) -> AlmanacResult<AzElRange>

Computes the azimuth (in degrees), elevation (in degrees), and range (in kilometers) of the receiver state (rx) seen from the location ID (as transmitter state, once converted into the SEZ frame of the transmitter. Refer to [azimuth_elevation_range_sez] for algorithm details.

pub fn azimuth_elevation_range_sez_from_location_name( &self, rx: Orbit, location_name: &str, obstructing_body: Option<Frame>, ab_corr: Option<Aberration>, ) -> AlmanacResult<AzElRange>

Computes the azimuth (in degrees), elevation (in degrees), and range (in kilometers) of the receiver state (rx) seen from the location ID (as transmitter state, once converted into the SEZ frame of the transmitter. Refer to [azimuth_elevation_range_sez] for algorithm details.

pub fn azimuth_elevation_range_sez_from_location( &self, rx: Orbit, location: Location, obstructing_body: Option<Frame>, ab_corr: Option<Aberration>, ) -> AlmanacResult<AzElRange>

Computes the azimuth (in degrees), elevation (in degrees), and range (in kilometers) of the receiver state (rx) seen from the provided location (as transmitter state, once converted into the SEZ frame of the transmitter. Refer to [azimuth_elevation_range_sez] for algorithm details. Location terrain masks are always applied, i.e. if the terrain masks the object, all data is set to f64::NAN, unless specified otherwise in the Location.

impl Almanac

pub fn from_bpc(bpc: BPC) -> Self

pub fn with_bpc(self, bpc: BPC) -> Self

Loads a new Binary Planetary Constants (BPC) kernel into a new context, using the system time as the alias. If the time is not availble, then 0 TAI is used. This new context is needed to satisfy the unloading of files. In fact, to unload a file, simply let the newly loaded context drop out of scope and Rust will clean it up.

pub fn with_bpc_as(self, bpc: BPC, alias: Option<String>) -> Self

Loads a new Binary Planetary Constant (BPC) file into a new context, naming it with the provided alias or the current system time. To unload a file, call bpc_unload.

pub fn bpc_unload(&mut self, alias: &str) -> Result<(), OrientationError>

Unloads the BPC with the provided alias. WARNING: This causes the order of the loaded files to be perturbed, which may be an issue if several SPKs with the same IDs are loaded.

pub fn num_loaded_bpc(&self) -> usize

pub fn bpc_summary_from_name_at_epoch( &self, name: &str, epoch: Epoch, ) -> Result<(&BPCSummaryRecord, usize, usize), OrientationError>

Returns the summary given the name of the summary record if that summary has data defined at the requested epoch and the BPC where this name was found to be valid at that epoch.

pub fn bpc_summary_at_epoch( &self, id: i32, epoch: Epoch, ) -> Result<(&BPCSummaryRecord, usize, usize), OrientationError>

Returns the summary given the name of the summary record if that summary has data defined at the requested epoch

pub fn bpc_summary_from_name( &self, name: &str, ) -> Result<(&BPCSummaryRecord, usize, usize), OrientationError>

Returns the summary given the name of the summary record.

pub fn bpc_summary( &self, id: i32, ) -> Result<(&BPCSummaryRecord, usize, usize), OrientationError>

Returns the summary given the name of the summary record if that summary has data defined at the requested epoch

impl Almanac

pub fn bpc_summaries( &self, id: i32, ) -> Result<Vec<BPCSummaryRecord>, OrientationError>

Returns a vector of the summaries whose ID matches the desired id, in the order in which they will be used, i.e. in reverse loading order.

Warning

This function performs a memory allocation.

:type id: int :rtype: typing.List

pub fn bpc_domain(&self, id: i32) -> Result<(Epoch, Epoch), OrientationError>

Returns the applicable domain of the request id, i.e. start and end epoch that the provided id has loaded data.

:type id: int :rtype: typing.Tuple

pub fn bpc_domains( &self, ) -> Result<HashMap<i32, (Epoch, Epoch)>, OrientationError>

Returns a map of each loaded BPC ID to its domain validity.

Warning

This function performs a memory allocation.

:rtype: typing.Dict

impl Almanac

pub fn line_of_sight_obstructed( &self, observer: Orbit, observed: Orbit, obstructing_body: Frame, ab_corr: Option<Aberration>, ) -> AlmanacResult<bool>

Computes whether the line of sight between an observer and an observed Cartesian state is obstructed by the obstructing body. Returns true if the obstructing body is in the way, false otherwise.

For example, if the Moon is in between a Lunar orbiter (observed) and a ground station (observer), then this function returns true because the Moon (obstructing body) is indeed obstructing the line of sight.

Observed
  o  -
   +    -
    +      -
     + ***   -
    * +    *   -
    *  + + * + + o
    *     *     Observer
      ****

Key Elements:

o represents the positions of the observer and observed objects.

The dashed line connecting the observer and observed is the line of sight.

Algorithm (source: Algorithm 35 of Vallado, 4th edition, page 308.):

r1 and r2 are the transformed radii of the observed and observer objects, respectively.

r1sq and r2sq are the squared magnitudes of these vectors.

r1dotr2 is the dot product of r1 and r2.

tau is a parameter that determines the intersection point along the line of sight.

The condition (1.0 - tau) * r1sq + r1dotr2 * tau <= ob_mean_eq_radius_km^2 checks if the line of sight is within the obstructing body’s radius, indicating an obstruction.

pub fn occultation( &self, back_frame: Frame, front_frame: Frame, observer: Orbit, ab_corr: Option<Aberration>, ) -> AlmanacResult<Occultation>

Computes the occultation percentage of the back_frame object by the front_frame object as seen from the observer, when according for the provided aberration correction.

A zero percent occultation means that the back object is fully visible from the observer. A 100% percent occultation means that the back object is fully hidden from the observer because of the front frame (i.e. umbra if the back object is the Sun). A value in between means that the back object is partially hidden from the observser (i.e. penumbra if the back object is the Sun). Refer to the MathSpec for modeling details.

pub fn solar_eclipsing( &self, eclipsing_frame: Frame, observer: Orbit, ab_corr: Option<Aberration>, ) -> AlmanacResult<Occultation>

Computes the solar eclipsing of the observer due to the eclipsing_frame.

This function calls occultation where the back object is the Sun in the J2000 frame, and the front object is the provided eclipsing frame.

:type eclipsing_frame: Frame :type observer: Orbit :type ab_corr: Aberration, optional :rtype: Occultation

pub fn beta_angle_deg( &self, state: Orbit, ab_corr: Option<Aberration>, ) -> AlmanacResult<f64>

Computes the Beta angle (β) for a given orbital state, in degrees. A Beta angle of 0° indicates that the orbit plane is edge-on to the Sun, leading to maximum eclipse time. Conversely, a Beta angle of +90° or -90° means the orbit plane is face-on to the Sun, resulting in continuous sunlight exposure and no eclipses.

The Beta angle (β) is defined as the angle between the orbit plane of a spacecraft and the vector from the central body (e.g., Earth) to the Sun. In simpler terms, it measures how much of the time a satellite in orbit is exposed to direct sunlight. The mathematical formula for the Beta angle is: β=arcsin(h⋅usun​) Where:

h is the unit vector of the orbital momentum.

usun​ is the unit vector pointing from the central body to the Sun.

Original code from GMAT, https://github.com/ChristopherRabotin/GMAT/blob/GMAT-R2022a/src/gmatutil/util/CalculationUtilities.cpp#L209-L219

pub fn local_solar_time( &self, state: Orbit, ab_corr: Option<Aberration>, ) -> AlmanacResult<Duration>

Compute the local solar time, returned as a Duration between 0 and 24 hours.

pub fn ltan( &self, orbit: Orbit, ab_corr: Option<Aberration>, ) -> AlmanacResult<Duration>

Returns the Local Time of the Ascending Node (LTAN). This is the local time on the celestial body at which the spacecraft crosses the equator from south to north.

The formula is from Wertz, “Spacecraft Attitude Determination and Control”, page 79, equation (4-8). LTAN (hours) = 12.0 + (RAAN_orbit - RA_sun) / 15.0

:type orbit: Orbit :type ab_corr: Aberration, optional :rtype: Duration

pub fn ltdn( &self, orbit: Orbit, ab_corr: Option<Aberration>, ) -> AlmanacResult<Duration>

Returns the Local Time of the Descending Node (LTDN) in hours. This is the local time on the celestial body at which the spacecraft crosses the equator from north to south.

LTDN is 12 hours after LTAN.

:type orbit: Orbit :type ab_corr: Aberration, optional :rtype: Duration

impl Almanac

pub fn frame_from_uid<U: Into<FrameUid>>( &self, uid: U, ) -> Result<Frame, PlanetaryDataError>

👎Deprecated since 0.7.0: use frame_info instead

Given the frame UID (or something that can be transformed into it), attempt to retrieve the full frame information, if that frame is loaded

pub fn frame_info<U: Into<FrameUid>>( &self, uid: U, ) -> Result<Frame, PlanetaryDataError>

Given the frame UID (or something that can be transformed into it), attempt to retrieve the full frame information, if that frame is loaded

pub fn with_planetary_data(&self, planetary_data: PlanetaryDataSet) -> Self

Loads the provided planetary data into a clone of this original Almanac.

impl Almanac

pub fn sun_angle_deg( &self, target_id: i32, observer_id: i32, epoch: Epoch, ab_corr: Option<Aberration>, ) -> Result<f64, EphemerisError>

Returns the angle (between 0 and 180 degrees) between the observer and the Sun, and the observer and the target body ID. This computes the Sun Probe Earth angle (SPE) if the probe is in a loaded SPK, its ID is the “observer_id”, and the target is set to its central body.

Geometry

If the SPE is greater than 90 degrees, then the celestial object below the probe is in sunlight.

Sunrise at nadir

Sun
 |  \      
 |   \
 |    \
 Obs. -- Target

Sun high at nadir

Sun
 \        
  \  __ θ > 90
   \     \
    Obs. ---------- Target

Sunset at nadir

         Sun
       /  
      /  __ θ < 90
     /    /
 Obs. -- Target

Algorithm

Compute the position of the Sun as seen from the observer

Compute the position of the target as seen from the observer

Return the arccosine of the dot product of the norms of these vectors.

:type target_id: int :type observer_id: int :type epoch: Epoch :type ab_corr: Aberration :rtype: float

pub fn sun_angle_deg_from_frame( &self, target: Frame, observer: Frame, epoch: Epoch, ab_corr: Option<Aberration>, ) -> Result<f64, EphemerisError>

Convenience function that calls sun_angle_deg with the provided frames instead of the ephemeris ID.

:type target: Frame :type observer: Frame :type epoch: Epoch :type ab_corr: Aberration :rtype: float

impl Almanac

pub fn from_spk(spk: SPK) -> Self

pub fn with_spk(self, spk: SPK) -> Self

Loads a new SPK file into a new context, using the system time as the alias. If the time is not availble, then 0 TAI is used. This new context is needed to satisfy the unloading of files. In fact, to unload a file, simply let the newly loaded context drop out of scope and Rust will clean it up.

pub fn with_spk_as(self, spk: SPK, alias: Option<String>) -> Self

Loads a new SPK file into a new context, naming it with the provided alias, or the current system time if no alias is provided. To unload a file, call spk_unload.

pub fn spk_unload(&mut self, alias: &str) -> Result<(), EphemerisError>

Unloads the SPK with the provided alias. WARNING: This causes the order of the loaded files to be perturbed, which may be an issue if several SPKs with the same IDs are loaded.

impl Almanac

pub fn num_loaded_spk(&self) -> usize

pub fn spk_summary_from_name_at_epoch( &self, name: &str, epoch: Epoch, ) -> Result<(&SPKSummaryRecord, usize, usize), EphemerisError>

Returns the summary given the name of the summary record if that summary has data defined at the requested epoch and the SPK where this name was found to be valid at that epoch.

pub fn spk_summary_at_epoch( &self, id: i32, epoch: Epoch, ) -> Result<(&SPKSummaryRecord, usize, usize), EphemerisError>

Returns the summary given the name of the summary record if that summary has data defined at the requested epoch

pub fn spk_summary_from_name( &self, name: &str, ) -> Result<(&SPKSummaryRecord, usize, usize), EphemerisError>

Returns the most recently loaded summary by its name, if any with that ID are available

pub fn spk_summary( &self, id: i32, ) -> Result<(&SPKSummaryRecord, usize, usize), EphemerisError>

Returns the most recently loaded summary by its ID, if any with that ID are available

impl Almanac

pub fn spk_summaries( &self, id: i32, ) -> Result<Vec<SPKSummaryRecord>, EphemerisError>

Returns a vector of the summaries whose ID matches the desired id, in the order in which they will be used, i.e. in reverse loading order.

Warning

This function performs a memory allocation.

:type id: int :rtype: typing.List

pub fn spk_domain(&self, id: i32) -> Result<(Epoch, Epoch), EphemerisError>

Returns the applicable domain of the request id, i.e. start and end epoch that the provided id has loaded data.

:type id: int :rtype: typing.Tuple

pub fn spk_domains( &self, ) -> Result<HashMap<i32, (Epoch, Epoch)>, EphemerisError>

Returns a map of each loaded SPK ID to its domain validity.

Warning

This function performs a memory allocation.

:rtype: typing.Dict

impl Almanac

pub fn transform( &self, target_frame: Frame, observer_frame: Frame, epoch: Epoch, ab_corr: Option<Aberration>, ) -> AlmanacResult<CartesianState>

Returns the Cartesian state needed to transform the target_frame to the observer_frame.

SPICE Compatibility

This function is the SPICE equivalent of spkezr: spkezr(TARGET_ID, EPOCH_TDB_S, ORIENTATION_ID, ABERRATION, OBSERVER_ID) In ANISE, the TARGET_ID and ORIENTATION are provided in the first argument (TARGET_FRAME), as that frame includes BOTH the target ID and the orientation of that target. The EPOCH_TDB_S is the epoch in the TDB time system, which is computed in ANISE using Hifitime. THe ABERRATION is computed by providing the optional Aberration flag. Finally, the OBSERVER argument is replaced by OBSERVER_FRAME: if the OBSERVER_FRAME argument has the same orientation as the TARGET_FRAME, then this call will return exactly the same data as the spkerz SPICE call.

Note

The units will be those of the underlying ephemeris data (typically km and km/s)

pub fn transform_to( &self, state: CartesianState, observer_frame: Frame, ab_corr: Option<Aberration>, ) -> AlmanacResult<CartesianState>

Returns the provided state as seen from the observer frame, given the aberration.

pub fn state_of( &self, object: i32, observer: Frame, epoch: Epoch, ab_corr: Option<Aberration>, ) -> AlmanacResult<CartesianState>

Returns the Cartesian state of the object as seen from the provided observer frame (essentially spkezr).

Note

The units will be those of the underlying ephemeris data (typically km and km/s)

pub fn spk_ezr( &self, target: i32, epoch: Epoch, frame: i32, observer: i32, ab_corr: Option<Aberration>, ) -> AlmanacResult<CartesianState>

Alias fo SPICE’s spkezr where the inputs must be the NAIF IDs of the objects and frames with the caveat that the aberration is moved to the last positional argument.

pub fn transform_state_to( &self, position: Vector3, velocity: Vector3, from_frame: Frame, to_frame: Frame, epoch: Epoch, ab_corr: Option<Aberration>, distance_unit: LengthUnit, time_unit: TimeUnit, ) -> AlmanacResult<CartesianState>

Translates a state with its origin (to_frame) and given its units (distance_unit, time_unit), returns that state with respect to the requested frame

WARNING: This function only performs the translation and no rotation whatsoever. Use the transform_state_to function instead to include rotations.

pub fn unit_vector( &self, target_frame: Frame, observer_frame: Frame, epoch: Epoch, ab_corr: Option<Aberration>, ) -> AlmanacResult<Vector3>

Returns the unitary 3D vector between two Frames (solid bodies) at desired Epoch

pub fn sun_unit_vector( &self, epoch: Epoch, observer_frame: Frame, ab_corr: Option<Aberration>, ) -> AlmanacResult<Vector3>

Returns the unitary 3D vector between desired Frame (solid body) and the Sun at desired Epoch

pub fn earth_sun_unit_vector( &self, epoch: Epoch, ab_corr: Option<Aberration>, ) -> AlmanacResult<Vector3>

Returns the unitary 3D vector between Earth and Sun at desired Epoch.

impl Almanac

pub fn load_from_metafile( self, metafile: MetaFile, autodelete: bool, ) -> AlmanacResult<Self>

Load from the provided MetaFile, downloading it if necessary. Set autodelete to true to automatically delete lock files. Lock files are important in multi-threaded loads.

impl Almanac

pub fn new(path: &str) -> AlmanacResult<Self>

Initializes a new Almanac from the provided file path, guessing at the file type

pub fn with_spacecraft_data(self, spacecraft_data: SpacecraftDataSet) -> Self

Loads the provided spacecraft data into a clone of this original Almanac.

pub fn with_euler_parameters(self, ep_dataset: EulerParameterDataSet) -> Self

Loads the provided Euler parameter data into a clone of this original Almanac.

pub fn with_location_data(self, loc_dataset: LocationDataSet) -> Self

Loads the provided location data into a clone of this original Almanac.

pub fn load_from_bytes(self, bytes: Bytes) -> AlmanacResult<Self>

Loads the provides bytes as one of the data types supported in ANISE.

pub fn load(self, path: &str) -> AlmanacResult<Self>

Generic function that tries to load the provided path guessing to the file type.

pub fn describe( &self, spk: Option<bool>, bpc: Option<bool>, planetary: Option<bool>, eulerparams: Option<bool>, locations: Option<bool>, time_scale: Option<TimeScale>, round_time: Option<bool>, )

Pretty prints the description of this Almanac, showing everything by default. Default time scale is TDB. If any parameter is set to true, then nothing other than that will be printed.

impl Almanac

pub fn report_event_once( &self, state_spec: &StateSpec, event: &Event, start_epoch: Epoch, end_epoch: Epoch, ) -> Result<EventDetails, AnalysisError>

Find the exact state where the request event happens. The event function is expected to be monotone in the provided interval because we find the event using a Brent solver. This will only return one event within the provided bracket.

pub fn report_events( &self, state_spec: &StateSpec, event: &Event, start_epoch: Epoch, end_epoch: Epoch, heuristic: Option<Duration>, ) -> Result<Vec<EventDetails>, AnalysisError>

Report all of the states and event details where the provided event occurs.

Limitations

This method uses a Brent solver, provides a superlinearity convergence (Golden ratio rate). If the function that defines the event is not unimodal, the event finder may not converge correctly. After the Brent solver is used, this function will check the median gap between events. Assuming most events are periodic, any gap whose median repetition is greater than 125% will be slow searches. This typically finds all of the events … but it may also add duplicates! To prevent reporting duplicate events, the found events are deduplicated if the same event is found within 3 times the epoch precision. For example, if the epoch precision is 100 ms, if three events are “found” within 300 ms of each other then only one of these three is preserved.

Heuristic detail

The initial search step is 1% of the duration requested, if the heuristic is set to None. For example, if the trajectory is 100 days long, then we split the trajectory into 100 chunks of 1 day and see whether the event is in there. If the event happens twice or more times within 1% of the trajectory duration, only the one of such events will be found.

pub fn report_events_slow( &self, state_spec: &StateSpec, event: &Event, start_epoch: Epoch, end_epoch: Epoch, ) -> Result<Vec<EventDetails>, AnalysisError>

Slow approach to finding all of the events between two epochs. This will evaluate ALL epochs in between the two bounds. This approach is more robust, but more computationally demanding since it’s O(N).

pub fn report_event_arcs( &self, state_spec: &StateSpec, event: &Event, start_epoch: Epoch, end_epoch: Epoch, ) -> Result<Vec<EventArc>, AnalysisError>

Find all event arcs, i.e. the start and stop time of when a given event occurs. This function calls the memory and computationally intensive [report_events_slow] function.

impl Almanac

pub fn report_scalars( &self, report: &ReportScalars, time_series: TimeSeries, ) -> HashMap<Epoch, Result<HashMap<String, AnalysisResult<f64>>, AnalysisError>>

Report a set of scalar expressions, optionally with aliases, at a fixed time step defined in the TimeSeries.

impl Almanac

pub fn try_find_ephemeris_root(&self) -> Result<i32, EphemerisError>

Returns the root of all of the loaded ephemerides, typically this should be the Solar System Barycenter.

Algorithm

For each loaded SPK, iterated in reverse order (to mimic SPICE behavior)

For each summary record in each SPK, follow the ephemeris branch all the way up until the end of this SPK or until the SSB.

pub fn ephemeris_path_to_root( &self, source: Frame, epoch: Epoch, ) -> Result<(usize, [Option<i32>; 8]), EphemerisError>

Try to construct the path from the source frame all the way to the root ephemeris of this context.

pub fn common_ephemeris_path( &self, from_frame: Frame, to_frame: Frame, epoch: Epoch, ) -> Result<(usize, [Option<i32>; 8], i32), EphemerisError>

Returns the ephemeris path between two frames and the common node. This may return a DisjointRoots error if the frames do not share a common root, which is considered a file integrity error.

Example

If the “from” frame is Earth Barycenter whose path to the ANISE root is the following:

Solar System barycenter
╰─> Earth Moon Barycenter
    ╰─> Earth

And the “to” frame is Moon, whose path is:

Solar System barycenter
╰─> Earth Moon Barycenter
    ╰─> Moon
        ╰─> LRO

Then this function will return the path an array of hashes of up to MAX_TREE_DEPTH items. In this example, the array with the hashes of the “Earth Moon Barycenter” and “Moon”.

Note

A proper ANISE file should only have a single root and if two paths are empty, then they should be the same frame. If a DisjointRoots error is reported here, it means that the ANISE file is invalid.

Time complexity

This can likely be simplified as this as a time complexity of O(n×m) where n, m are the lengths of the paths from the ephemeris up to the root. This can probably be optimized to avoid rewinding the entire frame path up to the root frame

impl Almanac

pub fn translate_to_parent( &self, source: Frame, epoch: Epoch, ) -> Result<CartesianState, EphemerisError>

Performs the GEOMETRIC translation to the parent. Use translate_from_to for aberration.

:type source: Frame :type epoch: Epoch :rtype: Orbit

impl Almanac

pub fn translate( &self, target_frame: Frame, observer_frame: Frame, epoch: Epoch, ab_corr: Option<Aberration>, ) -> Result<CartesianState, EphemerisError>

Returns the Cartesian state of the target frame as seen from the observer frame at the provided epoch, and optionally given the aberration correction.

SPICE Compatibility

This function is the SPICE equivalent of spkezr: spkezr(TARGET_ID, EPOCH_TDB_S, ORIENTATION_ID, ABERRATION, OBSERVER_ID) In ANISE, the TARGET_ID and ORIENTATION are provided in the first argument (TARGET_FRAME), as that frame includes BOTH the target ID and the orientation of that target. The EPOCH_TDB_S is the epoch in the TDB time system, which is computed in ANISE using Hifitime. THe ABERRATION is computed by providing the optional Aberration flag. Finally, the OBSERVER argument is replaced by OBSERVER_FRAME: if the OBSERVER_FRAME argument has the same orientation as the TARGET_FRAME, then this call will return exactly the same data as the spkerz SPICE call.

Warning

This function only performs the translation and no rotation whatsoever. Use the transform function instead to include rotations.

Note

This function performs a recursion of no more than twice the MAX_TREE_DEPTH.

Algorithm

Find the common ancestor of the target_frame and observer_frame in the ephemeris tree using common_ephemeris_path.

Initialize the state vectors for both the forward (observer to common ancestor) and backward (target to common ancestor) paths.

Iteratively traverse the ephemeris tree from the observer and target frames up to the common ancestor, accumulating the state vectors at each step using translation_parts_to_parent.

If aberration corrections are requested, calculate the one-way light time and apply the correction to the target’s position.

The final state is the difference between the backward and forward state vectors.

pub fn translate_geometric( &self, target_frame: Frame, observer_frame: Frame, epoch: Epoch, ) -> Result<CartesianState, EphemerisError>

Returns the geometric position vector, velocity vector, and acceleration vector needed to translate the from_frame to the to_frame, where the distance is in km, the velocity in km/s, and the acceleration in km/s^2.

pub fn translate_to( &self, state: CartesianState, observer_frame: Frame, ab_corr: Option<Aberration>, ) -> Result<CartesianState, EphemerisError>

Translates the provided Cartesian state into the requested observer frame

WARNING: This function only performs the translation and no rotation whatsoever. Use the transform_to function instead to include rotations.

pub fn translate_state_to( &self, position: Vector3, velocity: Vector3, from_frame: Frame, observer_frame: Frame, epoch: Epoch, ab_corr: Option<Aberration>, distance_unit: LengthUnit, time_unit: TimeUnit, ) -> Result<CartesianState, EphemerisError>

Translates a state with its origin (to_frame) and given its units (distance_unit, time_unit), returns that state with respect to the requested frame

WARNING: This function only performs the translation and no rotation whatsoever. Use the transform_state_to function instead to include rotations.

impl Almanac

pub fn try_find_orientation_root(&self) -> Result<i32, OrientationError>

Returns the root of all of the loaded orientations (BPC or planetary), typically this should be J2000.

Algorithm

For each loaded BPC, iterated in reverse order (to mimic SPICE behavior)

For each summary record in each BPC, follow the orientation branch all the way up until the end of this BPC or until the J2000.

pub fn orientation_path_to_root( &self, source: Frame, epoch: Epoch, ) -> Result<(usize, [Option<i32>; 8]), OrientationError>

Try to construct the path from the source frame all the way to the root orientation of this context.

pub fn common_orientation_path( &self, from_frame: Frame, to_frame: Frame, epoch: Epoch, ) -> Result<(usize, [Option<i32>; 8], i32), OrientationError>

Returns the orientation path between two frames and the common node. This may return a DisjointRoots error if the frames do not share a common root, which is considered a file integrity error.

impl Almanac

pub fn rotation_to_parent( &self, source: Frame, epoch: Epoch, ) -> Result<DCM, OrientationError>

Returns the direct cosine matrix (DCM) to rotate from the source to its parent in the orientation hierarchy at the provided epoch,

Example

If the ephemeris stores position interpolation coefficients in kilometer but this function is called with millimeters as a distance unit, the output vectors will be in mm, mm/s, mm/s^2 respectively.

Errors

As of now, some interpolation types are not supported, and if that were to happen, this would return an error.

WARNING: This function only performs the rotation and no translation whatsoever. Use the transform_to_parent_from function instead to include rotations.

impl Almanac

pub fn rotate( &self, from_frame: Frame, to_frame: Frame, epoch: Epoch, ) -> Result<DCM, OrientationError>

Returns the 6x6 DCM needed to rotation the from_frame to the to_frame.

Warning

This function only performs the rotation and no translation whatsoever. Use the transform_from_to function instead to include rotations.

Note

This function performs a recursion of no more than twice the MAX_TREE_DEPTH.

Algorithm

Find the common ancestor of the from_frame and to_frame in the orientation tree using common_orientation_path.

Initialize the DCMs for both the forward (from to common ancestor) and backward (to to common ancestor) paths.

Iteratively traverse the orientation tree from the from and to frames up to the common ancestor, composing the DCMs at each step using rotation_to_parent.

The final DCM is the composition of the forward and backward DCMs.

pub fn rotate_to( &self, state: CartesianState, observer_frame: Frame, ) -> Result<CartesianState, OrientationError>

Rotates the provided Cartesian state into the requested observer frame

WARNING: This function only performs the translation and no rotation whatsoever. Use the transform_to function instead to include rotations.

pub fn angular_velocity_rad_s( &self, from_frame: Frame, to_frame: Frame, epoch: Epoch, ) -> Result<Vector3, OrientationError>

Returns the angular velocity vector in rad/s of the from_frame wtr to the to_frame.

This can be used to compute the angular velocity of the Earth ITRF93 frame with respect to the J2000 frame for example.

pub fn angular_velocity_wtr_j2000_rad_s( &self, from_frame: Frame, epoch: Epoch, ) -> Result<Vector3, OrientationError>

Returns the angular velocity vector in rad/s of the from_frame wtr to the J2000 frame.

pub fn angular_velocity_deg_s( &self, from_frame: Frame, to_frame: Frame, epoch: Epoch, ) -> Result<Vector3, OrientationError>

Returns the angular velocity vector in deg/s of the from_frame wtr to the to_frame.

This can be used to compute the angular velocity of the Earth ITRF93 frame with respect to the J2000 frame for example.

pub fn angular_velocity_wtr_j2000_deg_s( &self, from_frame: Frame, epoch: Epoch, ) -> Result<Vector3, OrientationError>

Returns the angular velocity vector in deg/s of the from_frame wtr to the J2000 frame.

pub fn rotate_state_to( &self, position: Vector3, velocity: Vector3, from_frame: Frame, to_frame: Frame, epoch: Epoch, distance_unit: LengthUnit, time_unit: TimeUnit, ) -> Result<CartesianState, OrientationError>

Rotates a state with its origin (to_frame) and given its units (distance_unit, time_unit), returns that state with respect to the requested frame

WARNING: This function only performs the translation and no rotation whatsoever. Use the transform_state_to function instead to include rotations.