anise::analysis::elements

Enum OrbitalElement

pub enum OrbitalElement {

    AoL,
    AoP,
    ApoapsisRadius,
    ApoapsisAltitude,
    C3,
    Declination,
    EccentricAnomaly,
    Eccentricity,
    Energy,
    FlightPathAngle,
    Height,
    Latitude,
    Longitude,
    Hmag,
    HX,
    HY,
    HZ,
    HyperbolicAnomaly,
    Inclination,
    MeanAnomaly,
    PeriapsisRadius,
    PeriapsisAltitude,
    Period,
    RightAscension,
    RAAN,
    Rmag,
    SemiParameter,
    SemiMajorAxis,
    SemiMinorAxis,
    TrueAnomaly,
    TrueLongitude,
    VelocityDeclination,
    Vmag,
    X,
    Y,
    Z,
    VX,
    VY,
    VZ,

}

Orbital element defines all of the supported orbital elements in ANISE, which are all built from a State.

Variants

AoL

Argument of Latitude (deg)

AoP

Argument of Periapse (deg)

ApoapsisRadius

Radius of apoapsis (km)

ApoapsisAltitude

Altitude of apoapsis (km)

C3

C_3 in (km/s)^2

Declination

Declination (deg) (also called elevation if in a body fixed frame)

EccentricAnomaly

Eccentric anomaly (deg)

Eccentricity

Eccentricity (no unit)

Energy

Specific energy

FlightPathAngle

Flight path angle (deg)

Height

Geodetic height (km)

Latitude

Geodetic latitude (deg)

Longitude

Geodetic longitude (deg)

Hmag

Orbital momentum

HX

X component of the orbital momentum vector

HY

Y component of the orbital momentum vector

HZ

Z component of the orbital momentum vector

HyperbolicAnomaly

Hyperbolic anomaly (deg), only valid for hyperbolic orbits

Inclination

Inclination (deg)

MeanAnomaly

Mean anomaly (deg)

PeriapsisRadius

Radius of periapse (km)

PeriapsisAltitude

Altitude of periapse (km)

Period

Orbital period (s)

RightAscension

Right ascension (deg)

RAAN

Right ascension of the ascending node (deg)

Rmag

Norm of the radius vector

SemiParameter

Semi parameter (km)

SemiMajorAxis

Semi major axis (km)

SemiMinorAxis

Semi minor axis (km)

TrueAnomaly

True anomaly

TrueLongitude

True longitude

VelocityDeclination

Velocity declination (deg)

Vmag

Norm of the velocity vector (km/s)

X

X component of the radius (km)

Y

Y component of the radius (km)

Z

Z component of the radius (km)

VX

X component of the velocity (km/s)

VY

Y component of the velocity (km/s)

VZ

Z component of the velocity (km/s)

Implementations

impl OrbitalElement

pub fn evaluate(self, orbit: Orbit) -> Result<f64, AnalysisError>

Computes this scalar expression for the provided orbit.

pub const fn is_angle(&self) -> bool

pub const fn unit(&self) -> &'static str