anise::analysis::event

Struct EventDetails

pub struct EventDetails {
    pub orbit: Orbit,
    pub edge: EventEdge,
    pub value: f64,
    pub prev_value: Option<f64>,
    pub next_value: Option<f64>,
    pub pm_duration: Duration,
    pub repr: String,
}

Represents the details of an event occurring along a trajectory.

EventDetails encapsulates the state at which a particular event occurs in a trajectory, along with additional information about the nature of the event. This struct is particularly useful for understanding the dynamics of the event, such as whether it represents a rising or falling edge, or if the edge is unclear.

Generics

S: Interpolatable - A type that represents the state of the trajectory. This type must implement the Interpolatable trait, ensuring that it can be interpolated and manipulated according to the trajectory’s requirements.

Fields
orbit: Orbit

The state of the trajectory at the found event. :rtype: Orbit

edge: EventEdge

Indicates whether the event is a rising edge, falling edge, or unclear. This helps in understanding the direction of change at the event point. :rtype: EventEdge

value: f64

Numerical evaluation of the event condition, e.g. if seeking the apoapsis, this returns the near zero :rtype: float

prev_value: Option<f64>

Numertical evaluation of the event condition one epoch step before the found event (used to compute the rising/falling edge). :rtype: float

next_value: Option<f64>

Numertical evaluation of the event condition one epoch step after the found event (used to compute the rising/falling edge). :rtype: float

pm_duration: Duration

Precision of the epoch for this value :rtype: Duration

repr: String

Store the representation of this event as a string because we can’t move or clone the event reference :rtype: str

Implementations

impl EventDetails

pub fn new( state: Orbit, value: f64, event: &Event, prev_state: Option<Orbit>, next_state: Option<Orbit>, almanac: &Almanac, ) -> Result<Self, AnalysisError>

Generates detailed information about an event at a specific epoch in a trajectory.

This takes an Epoch as an input and returns a Result<Self, EventError>. It is designed to determine the state of a trajectory at a given epoch, evaluate the specific event at that state, and ascertain the nature of the event (rising, falling, or unclear). The initialization intelligently determines the edge type of the event by comparing the event’s value at the current, previous, and next epochs. It ensures robust event characterization in trajectories.

Returns

Ok(EventDetails<S>) if the state at the given epoch can be determined and the event details are successfully evaluated.

Err(EventError) if there is an error in retrieving the state at the specified epoch.