anise::analysis::expr

Enum ScalarExpr

pub enum ScalarExpr {

    Constant(f64),
    MeanEquatorialRadius {
        celestial_object: i32,
    },
    SemiMajorEquatorialRadius {
        celestial_object: i32,
    },
    SemiMinorEquatorialRadius {
        celestial_object: i32,
    },
    PolarRadius {
        celestial_object: i32,
    },
    Flattening {
        celestial_object: i32,
    },
    GravParam {
        celestial_object: i32,
    },
    Add {
        a: Box<Self>,
        b: Box<Self>,
    },
    Mul {
        a: Box<Self>,
        b: Box<Self>,
    },
    Negate(Box<Self>),
    Invert(Box<Self>),
    Sqrt(Box<Self>),
    Powi {
        scalar: Box<Self>,
        n: i32,
    },
    Powf {
        scalar: Box<Self>,
        n: f64,
    },
    Cos(Box<Self>),
    Sin(Box<Self>),
    Tan(Box<Self>),
    Acos(Box<Self>),
    Asin(Box<Self>),
    Atan2 {
        y: Box<Self>,
        x: Box<Self>,
    },
    Modulo {
        v: Box<Self>,
        m: Box<Self>,
    },
    Norm(VectorExpr),
    NormSquared(VectorExpr),
    DotProduct {
        a: VectorExpr,
        b: VectorExpr,
    },
    AngleBetween {
        a: VectorExpr,
        b: VectorExpr,
    },
    VectorX(VectorExpr),
    VectorY(VectorExpr),
    VectorZ(VectorExpr),
    Element(OrbitalElement),
    SolarEclipsePercentage {
        eclipsing_frame: Frame,
    },
    OccultationPercentage {
        back_frame: Frame,
        front_frame: Frame,
    },
    BetaAngle,
    LocalSolarTime,
    LocalTimeAscNode,
    LocalTimeDescNode,
    SunAngle {
        observer_id: i32,
    },
    AzimuthFromLocation {
        location_id: i32,
        obstructing_body: Option<Frame>,
    },
    ElevationFromLocation {
        location_id: i32,
        obstructing_body: Option<Frame>,
    },
    RangeFromLocation {
        location_id: i32,
        obstructing_body: Option<Frame>,
    },
    RangeRateFromLocation {
        location_id: i32,
        obstructing_body: Option<Frame>,
    },

}

ScalarExpr defines a scalar computation from a (set of) vector expression(s).

Variants

Constant(f64)

MeanEquatorialRadius

Mean radius of the provided body, must be loaded in the almanac

Fields

celestial_object: i32

SemiMajorEquatorialRadius

Fields

celestial_object: i32

SemiMinorEquatorialRadius

Fields

celestial_object: i32

PolarRadius

Fields

celestial_object: i32

Flattening

Fields

celestial_object: i32

GravParam

Fields

celestial_object: i32

Add

Fields

a: Box<Self>

b: Box<Self>

Mul

Fields

a: Box<Self>

b: Box<Self>

Negate(Box<Self>)

Invert(Box<Self>)

Sqrt(Box<Self>)

Powi

Fields

scalar: Box<Self>

n: i32

Powf

Fields

scalar: Box<Self>

n: f64

Cos(Box<Self>)

Sin(Box<Self>)

Tan(Box<Self>)

Acos(Box<Self>)

Compute the arccos, returned in degrees

Asin(Box<Self>)

Compute the arcsin, returned in degrees

Atan2

Compute the arctan2 (i.e. arctan with quadrant check), returned in degrees

Fields

y: Box<Self>

x: Box<Self>

Modulo

Computes v % m

Fields

v: Box<Self>

m: Box<Self>

Norm(VectorExpr)

NormSquared(VectorExpr)

DotProduct

Fields

a: VectorExpr

b: VectorExpr

AngleBetween

Angle between two vectors, in degrees

Fields

a: VectorExpr

b: VectorExpr

VectorX(VectorExpr)

VectorY(VectorExpr)

VectorZ(VectorExpr)

Element(OrbitalElement)

SolarEclipsePercentage

Computes the eclipsing percentage due to the eclipsing frame. Aberration correction is that of the state spec.

Fields

eclipsing_frame: Frame

OccultationPercentage

Computes the occultation percentage of the back_frame due to the front_frame. Aberration correction is that of the state spec.

Fields

back_frame: Frame

front_frame: Frame

BetaAngle

Computes the beta angle, in degrees. Aberration correction is that of the state spec.

LocalSolarTime

Compute the local solar time, in hours

LocalTimeAscNode

Computes the local time of the ascending node, in hours

LocalTimeDescNode

Computes the local time of the descending node, in hours

SunAngle

Computes the Sun angle where observer_id is the ID of the spacecraft for example. If the frame of the state spec is in an Earth frame, then this computes the Sun Probe Earth angle. Refer to the sun_angle_deg function for detailed documentation.

Fields

observer_id: i32

AzimuthFromLocation

Fields

location_id: i32

obstructing_body: Option<Frame>

ElevationFromLocation

Fields

location_id: i32

obstructing_body: Option<Frame>

RangeFromLocation

Fields

location_id: i32

obstructing_body: Option<Frame>

RangeRateFromLocation

Fields

location_id: i32

obstructing_body: Option<Frame>

Implementations

impl ScalarExpr

pub fn evaluate( &self, orbit: Orbit, ab_corr: Option<Aberration>, almanac: &Almanac, ) -> Result<f64, AnalysisError>

Computes this scalar expression for the provided orbit.

pub fn to_s_expr(&self) -> Result<String, Error>

Export this Scalar Expression to S-Expression / LISP syntax

pub fn from_s_expr(expr: &str) -> Result<Self, Error>

Load this Scalar Expression from an S-Expression / LISP syntax