anise::astro::orbit

Type Alias Orbit

pub type Orbit = CartesianState;

A helper type alias, but no assumptions are made on the underlying validity of the frame.

Aliased Type

pub struct Orbit {
    pub radius_km: Matrix<f64, Const<3>, Const<1>, ArrayStorage<f64, 3, 1>>,
    pub velocity_km_s: Matrix<f64, Const<3>, Const<1>, ArrayStorage<f64, 3, 1>>,
    pub epoch: Epoch,
    pub frame: Frame,
}

Fields
radius_km: Matrix<f64, Const<3>, Const<1>, ArrayStorage<f64, 3, 1>>

Position radius in kilometers

velocity_km_s: Matrix<f64, Const<3>, Const<1>, ArrayStorage<f64, 3, 1>>

Velocity in kilometers per second

epoch: Epoch

Epoch with time scale at which this is valid.

frame: Frame

Frame in which this Cartesian state lives.

Implementations

impl Orbit

pub fn try_keplerian( sma_km: f64, ecc: f64, inc_deg: f64, raan_deg: f64, aop_deg: f64, ta_deg: f64, epoch: Epoch, frame: Frame, ) -> PhysicsResult<Self>

Attempts to create a new Orbit around the provided Celestial or Geoid frame from the Keplerian orbital elements.

Units: km, none, degrees, degrees, degrees, degrees

WARNING: This function will return an error if the singularities in the conversion are encountered. NOTE: The state is defined in Cartesian coordinates as they are non-singular. This causes rounding errors when creating a state from its Keplerian orbital elements (cf. the state tests). One should expect these errors to be on the order of 1e-12.

pub fn try_keplerian_apsis_radii( r_a_km: f64, r_p_km: f64, inc_deg: f64, raan_deg: f64, aop_deg: f64, ta_deg: f64, epoch: Epoch, frame: Frame, ) -> PhysicsResult<Self>

Attempts to create a new Orbit from the provided radii of apoapsis and periapsis, in kilometers

pub fn try_keplerian_vec( state: &Vector6, epoch: Epoch, frame: Frame, ) -> PhysicsResult<Self>

Attempts to create a new Orbit around the provided frame from the borrowed state vector

The state vector must be sma, ecc, inc, raan, aop, ta. This function is a shortcut to cartesian and as such it has the same unit requirements.

pub fn keplerian( sma_km: f64, ecc: f64, inc_deg: f64, raan_deg: f64, aop_deg: f64, ta_deg: f64, epoch: Epoch, frame: Frame, ) -> Self

Creates (without error checking) a new Orbit around the provided Celestial or Geoid frame from the Keplerian orbital elements.

Units: km, none, degrees, degrees, degrees, degrees

WARNING: This function will panic if the singularities in the conversion are expected. NOTE: The state is defined in Cartesian coordinates as they are non-singular. This causes rounding errors when creating a state from its Keplerian orbital elements (cf. the state tests). One should expect these errors to be on the order of 1e-12.

pub fn keplerian_apsis_radii( r_a_km: f64, r_p_km: f64, inc_deg: f64, raan_deg: f64, aop_deg: f64, ta_deg: f64, epoch: Epoch, frame: Frame, ) -> Self

Creates a new Orbit from the provided radii of apoapsis and periapsis, in kilometers

pub fn try_keplerian_mean_anomaly( sma_km: f64, ecc: f64, inc_deg: f64, raan_deg: f64, aop_deg: f64, ma_deg: f64, epoch: Epoch, frame: Frame, ) -> PhysicsResult<Self>

Initializes a new orbit from the Keplerian orbital elements using the mean anomaly instead of the true anomaly.

Implementation notes

This function starts by converting the mean anomaly to true anomaly, and then it initializes the orbit using the keplerian(..) method. The conversion is from GMAT’s MeanToTrueAnomaly function, transliterated originally by Claude and GPT4 with human adjustments.

pub fn keplerian_vec(state: &Vector6, epoch: Epoch, frame: Frame) -> Self

Creates a new Orbit around the provided frame from the borrowed state vector

The state vector must be sma, ecc, inc, raan, aop, ta. This function is a shortcut to cartesian and as such it has the same unit requirements.

pub fn to_keplerian_vec(self) -> PhysicsResult<Vector6>

Returns this state as a Keplerian Vector6 in [km, none, degrees, degrees, degrees, degrees]

Note that the time is not returned in the vector.

pub fn hvec(&self) -> PhysicsResult<Vector3>

Returns the orbital momentum vector

pub fn h_hat(&self) -> PhysicsResult<Vector3>

Returns the orbital momentum unit vector

pub fn evec(&self) -> Result<Vector3, PhysicsError>

Returns the eccentricity vector (no unit)

impl Orbit

pub fn dcm_from_topocentric_to_body_fixed( &self, _from: i32, ) -> PhysicsResult<DCM>

Builds the rotation matrix that rotates from the topocentric frame (SEZ) into the body fixed frame of this state.

Frame warnings

If the state is NOT in a body fixed frame (i.e. ITRF93), then this computation is INVALID.

(Usually) no time derivative can be computed: the orbit is expected to be a body fixed frame where the at_epoch function will fail. Exceptions for Moon body fixed frames.

UNUSED Arguments

from: ID of this new frame. Only used to set the “from” frame of the DCM. – No longer used since 0.5.3

Source

From the GMAT MathSpec, page 30 section 2.6.9 and from Calculate_RFT in TopocentricAxes.cpp, this returns the rotation matrix from the topocentric frame (SEZ) to body fixed frame. In the GMAT MathSpec notation, R_{IF} is the DCM from body fixed to inertial. Similarly, R{FT} is from topocentric to body fixed.

:type _from: int :rtype: DCM

pub fn dcm3x3_from_topocentric_to_body_fixed(&self) -> PhysicsResult<DCM>

Builds the rotation matrix that rotates from the topocentric frame (SEZ) into the body fixed frame of this state.

Frame warning

If the state is NOT in a body fixed frame (i.e. ITRF93), then this computation is INVALID.

Source

From the GMAT MathSpec, page 30 section 2.6.9 and from Calculate_RFT in TopocentricAxes.cpp, this returns the rotation matrix from the topocentric frame (SEZ) to body fixed frame. In the GMAT MathSpec notation, R_{IF} is the DCM from body fixed to inertial. Similarly, R{FT} is from topocentric to body fixed.

:rtype: DCM

pub fn dcm_from_ric_to_inertial(&self) -> PhysicsResult<DCM>

Builds the rotation matrix that rotates from this state’s inertial frame to this state’s RIC frame

Frame warning

If the state is NOT in an inertial frame, then this computation is INVALID.

Algorithm

Compute the state data one millisecond before and one millisecond assuming two body dynamics

Compute the DCM for this state, and the pre and post states

Build the c vector as the normalized orbital momentum vector

Build the i vector as the cross product of \hat{r} and c

Build the RIC DCM as a 3x3 of the columns [\hat{r}, \hat{i}, \hat{c}], for the post, post, and current states

Compute the difference between the DCMs of the pre and post states, to build the DCM time derivative

Return the DCM structure with a 6x6 state DCM.

Note on the time derivative

If the pre or post states cannot be computed, then the time derivative of the DCM will not be set. Further note that most astrodynamics tools do not account for the time derivative in the RIC frame.

:rtype: DCM

pub fn dcm3x3_from_ric_to_inertial(&self) -> PhysicsResult<DCM>

Builds the rotation matrix that rotates from this state’s inertial frame to this state’s RIC frame

Frame warning

If the state is NOT in an inertial frame, then this computation is INVALID.

Algorithm

Build the c vector as the normalized orbital momentum vector

Build the i vector as the cross product of \hat{r} and c

Build the RIC DCM as a 3x3 of the columns [\hat{r}, \hat{i}, \hat{c}]

Return the DCM structure without accounting for the transport theorem.

:rtype: DCM

pub fn dcm3x3_from_rcn_to_inertial(&self) -> PhysicsResult<DCM>

Builds the rotation matrix that rotates from this state’s inertial frame to this state’s RCN frame (radial, cross, normal)

Frame warning

If the stattion is NOT in an inertial frame, then this computation is INVALID.

Algorithm

Compute \hat{r}, \hat{h}, the unit vectors of the radius and orbital momentum.

Compute the cross product of these

Build the DCM with these unit vectors

Return the DCM structure

:rtype: DCM

pub fn dcm_from_rcn_to_inertial(&self) -> PhysicsResult<DCM>

Builds the rotation matrix that rotates from this state’s inertial frame to this state’s RCN frame (radial, cross, normal)

Frame warning

If the stattion is NOT in an inertial frame, then this computation is INVALID.

Algorithm

Compute \hat{r}, \hat{h}, the unit vectors of the radius and orbital momentum.

Compute the cross product of these

Build the DCM with these unit vectors

Return the DCM structure with a 6x6 DCM with the time derivative of the VNC frame set.

Note on the time derivative

If the pre or post states cannot be computed, then the time derivative of the DCM will not be set. Further note that most astrodynamics tools do not account for the time derivative in the RIC frame.

:rtype: DCM

pub fn dcm3x3_from_vnc_to_inertial(&self) -> PhysicsResult<DCM>

Builds the rotation matrix that rotates from this state’s inertial frame to this state’s VNC frame (velocity, normal, cross)

Frame warning

If the stattion is NOT in an inertial frame, then this computation is INVALID.

Algorithm

Compute \hat{v}, \hat{h}, the unit vectors of the radius and orbital momentum.

Compute the cross product of these

Build the DCM with these unit vectors

Return the DCM structure.

:rtype: DCM

pub fn dcm_from_vnc_to_inertial(&self) -> PhysicsResult<DCM>

Builds the rotation matrix that rotates from this state’s inertial frame to this state’s VNC frame (velocity, normal, cross)

Frame warning

If the stattion is NOT in an inertial frame, then this computation is INVALID.

Algorithm

Compute \hat{v}, \hat{h}, the unit vectors of the radius and orbital momentum.

Compute the cross product of these

Build the DCM with these unit vectors

Compute the difference between the DCMs of the pre and post states (+/- 1 ms), to build the DCM time derivative

Return the DCM structure with a 6x6 DCM with the time derivative of the VNC frame set.

Note on the time derivative

If the pre or post states cannot be computed, then the time derivative of the DCM will not be set. Further note that most astrodynamics tools do not account for the time derivative in the RIC frame.

:rtype: DCM

pub fn hx(&self) -> PhysicsResult<f64>

Returns the orbital momentum value on the X axis

:rtype: float

pub fn hy(&self) -> PhysicsResult<f64>

Returns the orbital momentum value on the Y axis

:rtype: float

pub fn hz(&self) -> PhysicsResult<f64>

Returns the orbital momentum value on the Z axis

:rtype: float

pub fn hmag(&self) -> PhysicsResult<f64>

Returns the norm of the orbital momentum

:rtype: float

pub fn energy_km2_s2(&self) -> PhysicsResult<f64>

Returns the specific mechanical energy in km^2/s^2

:rtype: float

pub fn sma_km(&self) -> PhysicsResult<f64>

Returns the semi-major axis in km

:rtype: float

pub fn set_sma_km(&mut self, new_sma_km: f64) -> PhysicsResult<()>

Mutates this orbit to change the SMA

:type new_sma_km: float :rtype: None

pub fn with_sma_km(&self, new_sma_km: f64) -> PhysicsResult<Self>

Returns a copy of the state with a new SMA

:type new_sma_km: float :rtype: Orbit

pub fn add_sma_km(&self, delta_sma_km: f64) -> PhysicsResult<Self>

Returns a copy of the state with a provided SMA added to the current one

:type delta_sma_km: float :rtype: Orbit

pub fn period(&self) -> PhysicsResult<Duration>

Returns the period in seconds

:rtype: Duration

pub fn mean_motion_deg_s(&self) -> PhysicsResult<f64>

Returns the mean motion in degrees per seconds

:rtype: float

pub fn ecc(&self) -> PhysicsResult<f64>

Returns the eccentricity (no unit)

:rtype: float

pub fn set_ecc(&mut self, new_ecc: f64) -> PhysicsResult<()>

Mutates this orbit to change the ECC

:type new_ecc: float :rtype: None

pub fn with_ecc(&self, new_ecc: f64) -> PhysicsResult<Self>

Returns a copy of the state with a new ECC

:type new_ecc: float :rtype: Orbit

pub fn add_ecc(&self, delta_ecc: f64) -> PhysicsResult<Self>

Returns a copy of the state with a provided ECC added to the current one

:type delta_ecc: float :rtype: Orbit

pub fn inc_deg(&self) -> PhysicsResult<f64>

Returns the inclination in degrees

:rtype: float

pub fn set_inc_deg(&mut self, new_inc_deg: f64) -> PhysicsResult<()>

Mutates this orbit to change the INC

:type new_inc_deg: float :rtype: None

pub fn with_inc_deg(&self, new_inc_deg: f64) -> PhysicsResult<Self>

Returns a copy of the state with a new INC

:type new_inc_deg: float :rtype: Orbit

pub fn add_inc_deg(&self, delta_inc_deg: f64) -> PhysicsResult<Self>

Returns a copy of the state with a provided INC added to the current one

:type delta_inc_deg: float :rtype: None

pub fn aop_deg(&self) -> PhysicsResult<f64>

Returns the argument of periapsis in degrees

:rtype: float

pub fn set_aop_deg(&mut self, new_aop_deg: f64) -> PhysicsResult<()>

Mutates this orbit to change the AOP

:type new_aop_deg: float :rtype: None

pub fn with_aop_deg(&self, new_aop_deg: f64) -> PhysicsResult<Self>

Returns a copy of the state with a new AOP

:type new_aop_deg: float :rtype: Orbit

pub fn add_aop_deg(&self, delta_aop_deg: f64) -> PhysicsResult<Self>

Returns a copy of the state with a provided AOP added to the current one

:type delta_aop_deg: float :rtype: Orbit

pub fn raan_deg(&self) -> PhysicsResult<f64>

Returns the right ascension of the ascending node in degrees

:rtype: float

pub fn set_raan_deg(&mut self, new_raan_deg: f64) -> PhysicsResult<()>

Mutates this orbit to change the RAAN

:type new_raan_deg: float :rtype: None

pub fn with_raan_deg(&self, new_raan_deg: f64) -> PhysicsResult<Self>

Returns a copy of the state with a new RAAN

:type new_raan_deg: float :rtype: Orbit

pub fn add_raan_deg(&self, delta_raan_deg: f64) -> PhysicsResult<Self>

Returns a copy of the state with a provided RAAN added to the current one

:type delta_raan_deg: float :rtype: Orbit

pub fn ta_deg(&self) -> PhysicsResult<f64>

Returns the true anomaly in degrees between 0 and 360.0

NOTE: This function will emit a warning stating that the TA should be avoided if in a very near circular orbit Code from https://github.com/ChristopherRabotin/GMAT/blob/80bde040e12946a61dae90d9fc3538f16df34190/src/gmatutil/util/StateConversionUtil.cpp#L6835

LIMITATION: For an orbit whose true anomaly is (very nearly) 0.0 or 180.0, this function may return either 0.0 or 180.0 with a very small time increment. This is due to the precision of the cosine calculation: if the arccosine calculation is out of bounds, the sign of the cosine of the true anomaly is used to determine whether the true anomaly should be 0.0 or 180.0. In other words, there is an ambiguity in the computation in the true anomaly exactly at 180.0 and 0.0.

:rtype: float

pub fn set_ta_deg(&mut self, new_ta_deg: f64) -> PhysicsResult<()>

Mutates this orbit to change the TA

:type new_ta_deg: float :rtype: None

pub fn ta_dot_deg_s(&self) -> PhysicsResult<f64>

Returns the time derivative of the true anomaly computed as the 360.0 degrees divided by the orbital period (in seconds).

:rtype: float

pub fn with_ta_deg(&self, new_ta_deg: f64) -> PhysicsResult<Self>

Returns a copy of the state with a new TA

:type new_ta_deg: float :rtype: Orbit

pub fn add_ta_deg(&self, delta_ta_deg: f64) -> PhysicsResult<Self>

Returns a copy of the state with a provided TA added to the current one

:type delta_ta_deg: float :rtype: Orbit

pub fn with_apoapsis_periapsis_km( &self, new_ra_km: f64, new_rp_km: f64, ) -> PhysicsResult<Self>

Returns a copy of this state with the provided apoasis and periapsis

:type new_ra_km: float :type new_rp_km: float :rtype: Orbit

pub fn add_apoapsis_periapsis_km( &self, delta_ra_km: f64, delta_rp_km: f64, ) -> PhysicsResult<Self>

Returns a copy of this state with the provided apoasis and periapsis added to the current values

:type delta_ra_km: float :type delta_rp_km: float :rtype: Orbit

pub fn tlong_deg(&self) -> PhysicsResult<f64>

Returns the true longitude in degrees

:rtype: float

pub fn aol_deg(&self) -> PhysicsResult<f64>

Returns the argument of latitude in degrees

NOTE: If the orbit is near circular, the AoL will be computed from the true longitude instead of relying on the ill-defined true anomaly.

:rtype: float

pub fn periapsis_km(&self) -> PhysicsResult<f64>

Returns the radius of periapsis (or perigee around Earth), in kilometers.

:rtype: float

pub fn apoapsis_km(&self) -> PhysicsResult<f64>

Returns the radius of apoapsis (or apogee around Earth), in kilometers.

:rtype: float

pub fn ea_deg(&self) -> PhysicsResult<f64>

Returns the eccentric anomaly in degrees

This is a conversion from GMAT’s StateConversionUtil::TrueToEccentricAnomaly

:rtype: float

pub fn fpa_deg(&self) -> PhysicsResult<f64>

Returns the flight path angle in degrees

:rtype: float

pub fn ma_deg(&self) -> PhysicsResult<f64>

Returns the mean anomaly in degrees

This is a conversion from GMAT’s StateConversionUtil::TrueToMeanAnomaly

:rtype: float

pub fn semi_parameter_km(&self) -> PhysicsResult<f64>

Returns the semi parameter (or semilatus rectum)

:rtype: float

pub fn is_brouwer_short_valid(&self) -> PhysicsResult<bool>

Returns whether this state satisfies the requirement to compute the Mean Brouwer Short orbital element set.

This is a conversion from GMAT’s StateConversionUtil::CartesianToBrouwerMeanShort. The details are at the log level info. NOTE: Mean Brouwer Short are only defined around Earth. However, nyx does not check the main celestial body around which the state is defined (GMAT does perform this verification).

:rtype: bool

pub fn right_ascension_deg(&self) -> f64

Returns the right ascension of this orbit in degrees

:rtype: float

pub fn declination_deg(&self) -> f64

Returns the declination of this orbit in degrees

:rtype: float

pub fn semi_minor_axis_km(&self) -> PhysicsResult<f64>

Returns the semi minor axis in km, includes code for a hyperbolic orbit

:rtype: float

pub fn velocity_declination_deg(&self) -> f64

Returns the velocity declination of this orbit in degrees

:rtype: float

pub fn c3_km2_s2(&self) -> PhysicsResult<f64>

Returns the $C_3$ of this orbit in km^2/s^2

:rtype: float

pub fn ltan_deg(&self) -> PhysicsResult<f64>

Returns the Longitude of the Ascending Node (LTAN), or an error of equatorial orbits

:rtype: float

pub fn vinf_periapsis_km(&self, turn_angle_degrees: f64) -> PhysicsResult<f64>

Returns the radius of periapse in kilometers for the provided turn angle of this hyperbolic orbit. Returns an error if the orbit is not hyperbolic.

:type turn_angle_degrees: float :rtype: float

pub fn vinf_turn_angle_deg(&self, periapsis_km: f64) -> PhysicsResult<f64>

Returns the turn angle in degrees for the provided radius of periapse passage of this hyperbolic orbit Returns an error if the orbit is not hyperbolic.

:type periapsis_km: float :rtype: float

pub fn hyperbolic_anomaly_deg(&self) -> PhysicsResult<f64>

Returns the hyperbolic anomaly in degrees between 0 and 360.0 Returns an error if the orbit is not hyperbolic.

:rtype: float

pub fn at_epoch(&self, new_epoch: Epoch) -> PhysicsResult<Self>

Adjusts the true anomaly of this orbit using the mean anomaly.

Astrodynamics note

This is not a true propagation of the orbit. This is akin to a two body propagation ONLY without any other force models applied. Use Nyx for high fidelity propagation.

:type new_epoch: Epoch :rtype: Orbit

pub fn duration_to_radius(&self, radius_km: f64) -> PhysicsResult<Duration>

Calculates the duration to reach a specific radius in the orbit.

This function computes the time it will take for the orbiting body to reach the given radius_km from its current position. The calculation assumes two-body dynamics and considers the direction of motion.

Assumptions & Limitations

Assumes pure Keplerian motion.

For elliptical orbits, if the radius is reachable at two points (ascending and descending parts of the orbit), this function calculates the time to reach the radius corresponding to the true anomaly in [0, PI] (typically the ascending part or up to apoapsis if starting past periapsis).

For circular orbits, if the radius is within the apoapse and periapse, then a duration of zero is returned.

For hyperbolic/parabolic orbits, the true anomaly at radius is also computed in [0, PI]. If this point is in the past, the function returns an error, as it doesn’t look for solutions on the departing leg if nu > PI would be required (unless current TA is already > PI and target radius is further along). The current implementation strictly uses the acos result, so nu_rad_at_radius is always 0 <= nu <= PI. This means it finds the time to reach the radius on the path from periapsis up to the point where true anomaly is PI.

:type radius_km: float :rtype: Duration

pub fn ric_difference(&self, other: &Self) -> PhysicsResult<Self>

Returns a Cartesian state representing the RIC difference between self and other, in position and velocity (with transport theorem). Refer to dcm_from_ric_to_inertial for details on the RIC frame.

Algorithm

Compute the RIC DCM of self

Rotate self into the RIC frame

Rotation other into the RIC frame

Compute the difference between these two states

Strip the astrodynamical information from the frame, enabling only computations from CartesianState

:type other: Orbit :rtype: Orbit

pub fn vnc_difference(&self, other: &Self) -> PhysicsResult<Self>

Returns a Cartesian state representing the VNC difference between self and other, in position and velocity (with transport theorem). Refer to dcm_from_vnc_to_inertial for details on the VNC frame.

Algorithm

Compute the VNC DCM of self

Rotate self into the VNC frame

Rotation other into the VNC frame

Compute the difference between these two states

Strip the astrodynamical information from the frame, enabling only computations from CartesianState

:type other: Orbit :rtype: Orbit