anise::constants

Module frames

Constants

EARTH_ECLIPJ2000

EARTH_ITRF93

High fidelity Earth centered body fixed frame by the NAIF, requires the “Earth high prec” BPC kernel

EARTH_J2000

EARTH_MOON_BARYCENTER_J2000

EME2000

IAU_EARTH_FRAME

Low fidelity Earth centered body fixed frame by the International Astronomical Union (IAU)

IAU_JUPITER_FRAME

IAU_MARS_FRAME

IAU_MERCURY_FRAME

Body fixed IAU rotation

IAU_MOON_FRAME

Low fidelity Moon centered body fixed frame by the International Astronomical Union (IAU)

IAU_NEPTUNE_FRAME

IAU_SATURN_FRAME

IAU_URANUS_FRAME

IAU_VENUS_FRAME

JUPITER_BARYCENTER_J2000

MARS_BARYCENTER_J2000

MARS_J2000

The Mars frame is not available in the standard DE files, and requires a Mars specific BSP.

MERCURY_J2000

MOON_J2000

MOON_ME_DE421_FRAME

High fidelity Moon Mean Earth equator body fixed frame (used for cartography), requires the Moon PA BPC kernel

MOON_ME_DE440_ME421_FRAME

High fidelity Moon Mean Earth equator body fixed frame of the DE440 (used for cartography), requires the Moon PA BPC kernel moon_pa_de440_200625.bpc.

MOON_ME_FRAME

High fidelity Moon Mean Earth equator body fixed frame (used for cartography), requires the Moon PA BPC kernel

MOON_PA_DE421_FRAME

High fidelity Moon Mean Earth equator body fixed frame (used for cartography), requires the Moon PA BPC kernel

MOON_PA_DE440_FRAME

High fidelity Moon Mean Earth equator body fixed frame (used for cartography), requires the Moon PA BPC kernel moon_pa_de440_200625. Note that the ID is the same as the MOON_PA_DE421.

MOON_PA_FRAME

High fidelity Moon Principal Axes body fixed frame (used for gravity field and mass concentrations), requires the Moon PA BPC kernel

NEPTUNE_BARYCENTER_J2000

PLUTO_BARYCENTER_J2000

SATURN_BARYCENTER_J2000

SSB_J2000

SUN_J2000

URANUS_BARYCENTER_J2000

VENUS_J2000