anise::constants

Module orientations

Defines the orientations known to ANISE and SPICE. References used in the constants. [1] Jay Lieske, ``Precession Matrix Based on IAU (1976) System of Astronomical Constants,‘’ Astron. Astrophys. 73, 282-284 (1979).

[2] E.M. Standish, Jr., ``Orientation of the JPL Ephemerides, DE 200/LE 200, to the Dynamical Equinox of J2000,‘’ Astron. Astrophys. 114, 297-302 (1982).

[3] E.M. Standish, Jr., ``Conversion of Ephemeris Coordinates from the B1950 System to the J2000 System,‘’ JPL IOM 314.6-581, 24 June 1985.

[4] E.M. Standish, Jr., ``The Equinox Offsets of the JPL Ephemeris,‘’ JPL IOM 314.6-929, 26 February 1988.

[5] Jay Lieske, ``Expressions for the Precession Quantities Based upon the IAU (1976) System of Astronomical Constants’’ Astron. Astrophys. 58, 1-16 (1977).

[6] Laura Bass and Robert Cesarone “Mars Observer Planetary Constants and Models” JPL D-3444 November 1990.

[7] “Explanatory Supplement to the Astronomical Almanac” edited by P. Kenneth Seidelmann. University Science Books, 20 Edgehill Road, Mill Valley, CA 94941 (1992)

Constants

B1950

Earth mean equator, dynamical equinox of B1950. The B1950 reference frame is obtained by precessing the J2000 frame backwards from Julian year 2000 to Besselian year 1950, using the 1976 IAU precession model. The rotation from B1950 to J2000 is [ -z ] [ theta ] [ -zeta ] 3 2 3 The values for z, theta, and zeta are computed from the formulas given in table 5 of [5]. z = 1153.04066200330“ theta = 1002.26108439117“ zeta = 1152.84248596724“

DE096

DE102

DE108

DE111

DE114

DE118

JPL Developmental Ephemeris (118). The DE-118 reference frame is nearly identical to the FK4 frame. It is also derived from the B1950 frame. Only the offset is different

DE122

DE125

DE130

DE140

JPL Developmental Ephemeris. (140) The DE-140 frame is the DE-400 frame rotated:

DE142

JPL Developmental Ephemeris. (142) The DE-142 frame is the DE-402 frame rotated:

DE143

JPL Developmental Ephemeris. (143) The DE-143 frame is the DE-403 frame rotated:

DE200

DE202

ECLIPB1950

Ecliptic coordinates based upon the B1950 frame. The value for the obliquity of the ecliptic at B1950 is taken from page 171 of [7].

ECLIPJ2000

Ecliptic coordinates based upon the J2000 frame. The value for the obliquity of the ecliptic at J2000 is taken from page 114 of [7] equation 3.222-1. This agrees with the expression given in [5].

FK4

Fundamental Catalog (4). The FK4 reference frame is derived from the B1950 frame by applying the equinox offset determined by Fricke. [ 0.525“ ] 3

GALACTIC

Galactic System II. The Galactic System II reference frame is defined by the following rotations: o o o [ 327 ] [ 62.6 ] [ 282.25 ] 3 1 3 In the absence of better information, we assume the rotations are relative to the FK4 frame.

IAU_EARTH

Low fidelity Earth frame orientation by the International Astronomical Union (IAU)

IAU_JUPITER

IAU_MARS

IAU_MERCURY

Body fixed IAU rotation

IAU_MOON

Low fidelity Moon frame orientation by the International Astronomical Union (IAU)

IAU_NEPTUNE

IAU_SATURN

IAU_URANUS

IAU_VENUS

ITRF93

High fidelity Earth frame orientation by the NAIF, requires the “Earth high prec” BPC kernel

J2000

Earth mean equator, dynamical equinox of J2000. The root reference frame for SPICE.

J2000_TO_ECLIPJ2000_ANGLE_RAD

Angle between J2000 to solar system ecliptic J2000 (ECLIPJ2000), in radians (about 23.43929 degrees). Apply this rotation about the X axis (R1)

MARSIAU

Mars Mean Equator and IAU vector of J2000. The IAU-vector at Mars is the point on the mean equator of Mars where the equator ascends through the earth mean equator. This vector is the cross product of Earth mean north with Mars mean north.

MOON_ME

High fidelity Moon Mean Earth equator orientation frame (used for cartography), requires the Moon PA BPC kernel

MOON_ME_DE421

High fidelity Moon Mean Earth orientation frame of the DE421 (used for cartography), requires the Moon PA BPC kernel

MOON_ME_DE440_ME421

High fidelity Moon Mean Earth orientation frame of the DE440 (used for cartography), requires the Moon PA BPC kernel moon_pa_de440_200625.bpc.

MOON_PA

High fidelity Moon Principal Axes orientation frame (used for gravity field and mass concentrations), requires the Moon PA BPC kernel

MOON_PA_DE421

High fidelity Moon Principal Axes orientation frame of the DE421 (used for gravity field and mass concentrations), requires the Moon PA BPC kernel

MOON_PA_DE440

High fidelity Moon Principal Axes orientation frame of the DE440 (used for gravity field and mass concentrations), requires the Moon PA BPC kernel moon_pa_de440_200625. Note that the ID is the same as the MOON_PA_DE421.

Functions

id_from_orientation_name

Converts the provided ID to its human name. Only works for the common celestial bodies. Should be compatible with CCSDS OEM names

id_to_orientation_nameDeprecated

orientation_name_from_id

Given the frame ID, try to return a human name Source: https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/req/frames.html#Appendix.%20%60%60Built%20in''%20Inertial%20Reference%20Frames