anise::math

Module rotation

Structs

DCM

Defines a direction cosine matrix from one frame ID to another frame ID, optionally with its time derivative. It provides a number of run-time checks that prevent invalid rotations.

EulerParameter

Represents the orientation of a rigid body in three-dimensional space using Euler parameters.

MRP

Represents the orientation of a rigid body in three-dimensional space using Modified Rodrigues Parameters (MRP).

Traits

Rotation

Functions

r1

Build a 3x3 rotation matrix around the X axis

r2

Build a 3x3 rotation matrix around the Y axis

r3

Build a 3x3 rotation matrix around the Z axis

r1_dot

Build the derivative of the 3x3 rotation matrix around the X axis

r2_dot

Build the derivative of the 3x3 rotation matrix around the Y axis

r3_dot

Build the derivative of the 3x3 rotation matrix around the Z axis

Type Aliases

Quaternion

Quaternion will always be a unit quaternion in ANISE, cf. EulerParameter.