anise::math::rotation

Struct DCM

pub struct DCM {
    pub rot_mat: Matrix3,
    pub rot_mat_dt: Option<Matrix3>,
    pub from: i32,
    pub to: i32,
}

Defines a direction cosine matrix from one frame ID to another frame ID, optionally with its time derivative. It provides a number of run-time checks that prevent invalid rotations.

:type np_rot_mat: numpy.array :type from_id: int :type to_id: int :type np_rot_mat_dt: numpy.array, optional :rtype: DCM

Fields
rot_mat: Matrix3

The rotation matrix itself

rot_mat_dt: Option<Matrix3>

The time derivative of the rotation matrix

from: i32

The source frame

to: i32

The destination frame

Implementations

impl DCM

pub fn r1(angle_rad: f64, from: i32, to: i32) -> Self

Returns a rotation matrix for a rotation about the X axis.

Source: euler1 function from Baslisk

Arguments

angle_rad - The angle of rotation in radians.

pub fn r2(angle_rad: f64, from: i32, to: i32) -> Self

Returns a rotation matrix for a rotation about the Y axis.

Source: euler2 function from Basilisk

Arguments

angle - The angle of rotation in radians.

pub fn r3(angle_rad: f64, from: i32, to: i32) -> Self

Returns a rotation matrix for a rotation about the Z axis.

Source: euler3 function from Basilisk

Arguments

angle_rad - The angle of rotation in radians.

pub fn state_dcm(&self) -> Matrix6

Returns the 6x6 DCM to rotate a state. If the time derivative of this DCM is defined, this 6x6 accounts for the transport theorem.

pub fn identity(from: i32, to: i32) -> Self

Builds an identity rotation

pub fn skew_symmetric(&self) -> Option<Matrix3>

Returns the skew symmetric matrix if this DCM defines a rotation rate.

pub fn angular_velocity_rad_s(&self) -> Option<Vector3>

Returns the angular velocity vector in deg/s of this DCM is it has a defined rotation rate.

pub fn angular_velocity_deg_s(&self) -> Option<Vector3>

Returns the angular velocity vector in deg/s if a rotation rate is defined.

impl DCM

pub fn transpose(&self) -> Self

Returns the transpose of this DCM

:rtype: DCM

pub fn is_identity(&self) -> bool

Returns whether this rotation is identity, checking first the frames and then the rotation matrix (but ignores its time derivative)

:rtype: bool

pub fn is_valid(&self, unit_tol: f64, det_tol: f64) -> bool

Returns whether the rot_mat of this DCM is a valid rotation matrix. The criteria for validity are: – The columns of the matrix are unit vectors, within a specified tolerance (unit_tol). – The determinant of the matrix formed by unitizing the columns of the input matrix is 1, within a specified tolerance. This criterion ensures that the columns of the matrix are nearly orthogonal, and that they form a right-handed basis (det_tol). Source: SPICE’s rotation.req

:type unit_tol: float :type det_tol: float :rtype: bool