anise::math::rotation

Struct MRP

pub struct MRP {
    pub s0: f64,
    pub s1: f64,
    pub s2: f64,
    pub from: i32,
    pub to: i32,
}

Represents the orientation of a rigid body in three-dimensional space using Modified Rodrigues Parameters (MRP).

Modified Rodrigues Parameters (MRP) are a set of three parameters s0, s1, and s2 used for representing the attitude of a rigid body. They are an alternative to quaternions (Euler parameters) and are particularly useful for avoiding singularities and providing a smooth representation of rotations.

MRPs are related to the quaternion parameters q by: s0 = q1 / (1 + q0) s1 = q2 / (1 + q0) s2 = q3 / (1 + q0)

Shadow MRP

For any rotation, there are two sets of MRPs that can represent it: the regular MRPs and the shadow MRPs. The shadow MRP is defined as -s / (s0^2 + s1^2 + s2^2), where s is the vector of MRPs [s0, s1, s2].

Applications

In the context of spacecraft attitude determination and control, MRPs are often used because they provide a numerically stable way to represent the attitude of a spacecraft without the singularities that are present with Euler angles and with a minimal set of parameters.

Fields
s0: f64s1: f64s2: f64from: i32to: i32
Implementations

impl MRP

pub fn new(s0: f64, s1: f64, s2: f64, from: i32, to: i32) -> Self

Creates a new normalized MRP

pub fn shadow(&self) -> Result<Self, MathError>

Computes the shadow MRP.

Returns

The shadow MRP as a new instance of MRP.

pub fn is_singular(&self) -> bool

Returns whether this MRP is singular.

pub fn normalize(&self) -> Self

If the norm of this MRP is greater than max_norm then this MRP is set to its shadow set

pub fn uvec_angle(&self) -> (Vector3, f64)

Returns the principal line of rotation (a unit vector) and the angle of rotation in radians

Note

If the MRP is singular, this returns an angle of zero and a vector of zero.

pub fn b_matrix(&self) -> Matrix3

Returns the 3x3 matrix which relates the body angular velocity vector w to the derivative of this MRP. dQ/dt = 1/4 [B(Q)] w

pub fn diff_eq(&self, w: Vector3) -> MRP

Returns the MRP derivative for this MRP and body angular velocity vector w. dQ/dt = 1/4 [B(Q)] w

pub fn relative_to(&self, rhs: &Self) -> Result<Self, PhysicsError>

Returns the relative MRP between self and the rhs MRP.