anise::naif::daf::daf

Struct GenericDAF

pub struct GenericDAF<R: NAIFSummaryRecord, W: MutKind> {
    pub bytes: W,
    pub crc32_checksum: u32,
    pub _daf_type: PhantomData<R>,
}

Fields
bytes: Wcrc32_checksum: u32_daf_type: PhantomData<R>
Implementations

impl<R: NAIFSummaryRecord, W: MutKind> GenericDAF<R, W>

pub fn crc32(&self) -> u32

Compute the CRC32 of the underlying bytes

pub fn scrub(&self) -> Result<(), IntegrityError>

Scrubs the data by computing the CRC32 of the bytes and making sure that it still matches the previously known hash

pub fn file_record(&self) -> Result<FileRecord, DAFError>

Reads and parses the file record from the DAF bytes. The file record is always the first 1024 bytes of the file.

pub fn name_record(&self) -> Result<NameRecord, DAFError>

Reads and parses the name record from the DAF bytes. The file record contains a pointer to the start of the name record.

pub fn daf_summary(&self) -> Result<SummaryRecord, DAFError>

Reads and parses the DAF summary record.

pub fn data_summaries(&self) -> Result<&[R], DAFError>

Parses and returns a slice of the data summaries. The summaries are located in the same record as the DAF summary.

pub fn summary_from_name(&self, name: &str) -> Result<(&R, usize), DAFError>

Returns the summary given the name of the summary record

pub fn summary_from_name_at_epoch( &self, name: &str, epoch: Epoch, ) -> Result<(&R, usize), DAFError>

Returns the summary given the name of the summary record if that summary has data defined at the requested epoch

pub fn summary_from_id(&self, id: i32) -> Result<(&R, usize), DAFError>

Returns the summary given the id of the summary record

pub fn summary_from_id_at_epoch( &self, id: i32, epoch: Epoch, ) -> Result<(&R, usize), DAFError>

Returns the summary given the name of the summary record if that summary has data defined at the requested epoch

pub fn data_from_name<'a, S: NAIFDataSet<'a>>( &'a self, name: &str, ) -> Result<S, DAFError>

Provided a name that is in the summary, return its full data, if name is available.

pub fn nth_data<'a, S: NAIFDataSet<'a>>( &'a self, idx: usize, ) -> Result<S, DAFError>

Provided a name that is in the summary, return its full data, if name is available. This function retrieves the data associated with the nth summary record.

pub fn comments(&self) -> Result<Option<String>, DAFError>

pub fn persist<P: AsRef<Path>>(&self, path: P) -> IoResult<()>

Writes the contents of this DAF file to a new location.

impl<R: NAIFSummaryRecord> GenericDAF<R, Bytes>

pub fn parse<B: Deref<Target = [u8]>>(bytes: B) -> Result<Self, DAFError>

Parse the provided bytes as a SPICE Double Array File.

DAF File Structure

A DAF is composed of three main parts:

File Record: The first record (1024 bytes) of the file. It contains metadata about the file, such as the endianness, the number of records, and pointers to the other sections.

Comment Area: An optional area for storing comments.

Summary/Name Records and Data: The remaining records contain the summary records, name records, and the actual data arrays. The file record contains pointers to the start of these sections.

Parsing Process

The entire file is read into a Bytes object.

The CRC32 checksum of the bytes is computed.

The file_record and name_record are parsed to ensure the file is a valid DAF.

pub fn check_then_parse<B: Deref<Target = [u8]>>( bytes: B, expected: u32, ) -> Result<Self, DAFError>

Parse the DAF only if the CRC32 checksum of the data is valid

pub fn load(path: &str) -> Result<Self, DAFError>

pub fn from_static<B: Deref<Target = [u8]>>( bytes: &'static B, ) -> Result<Self, DAFError>

Parse the provided static byte array as a SPICE Double Array File

pub fn to_mutable(&self) -> MutDAF<R>

Copies the underlying bytes of this DAF into a MutDAF, enabling modification of the DAF.

impl<R: NAIFSummaryRecord> GenericDAF<R, BytesMut>

pub fn parse<B: Deref<Target = [u8]>>(bytes: B) -> Result<Self, DAFError>

Parse the provided bytes as a SPICE Double Array File

pub fn load(path: &str) -> Result<Self, DAFError>

pub fn set_name_record( &mut self, new_name_record: NameRecord, ) -> Result<(), DAFError>

Sets the name record of this mutable DAF file to the one provided as a parameter.

pub fn set_nth_data<'a, S: NAIFDataSet<'a>>( &mut self, idx: usize, new_data: S, new_start_epoch: Epoch, new_end_epoch: Epoch, ) -> Result<(), DAFError>

Sets the data for the n-th segment of this DAF file.

pub fn delete_nth_data(&mut self, idx: usize) -> Result<(), DAFError>

Deletes the data for the n-th segment of this DAF file.