anise::naif::daf::daf

Type Alias DAF

pub type DAF<R> = GenericDAF<R, Bytes>;

Aliased Type

pub struct DAF<R> {
    pub bytes: Bytes,
    pub crc32_checksum: u32,
    pub _daf_type: PhantomData<R>,
}

Fields
bytes: Bytescrc32_checksum: u32_daf_type: PhantomData<R>
Implementations

impl<R: NAIFSummaryRecord> DAF<R>

pub fn parse<B: Deref<Target = [u8]>>(bytes: B) -> Result<Self, DAFError>

Parse the provided bytes as a SPICE Double Array File.

DAF File Structure

A DAF is composed of three main parts:

File Record: The first record (1024 bytes) of the file. It contains metadata about the file, such as the endianness, the number of records, and pointers to the other sections.

Comment Area: An optional area for storing comments.

Summary/Name Records and Data: The remaining records contain the summary records, name records, and the actual data arrays. The file record contains pointers to the start of these sections.

Parsing Process

The entire file is read into a Bytes object.

The CRC32 checksum of the bytes is computed.

The file_record and name_record are parsed to ensure the file is a valid DAF.

pub fn check_then_parse<B: Deref<Target = [u8]>>( bytes: B, expected: u32, ) -> Result<Self, DAFError>

Parse the DAF only if the CRC32 checksum of the data is valid

pub fn load(path: &str) -> Result<Self, DAFError>

pub fn from_static<B: Deref<Target = [u8]>>( bytes: &'static B, ) -> Result<Self, DAFError>

Parse the provided static byte array as a SPICE Double Array File

pub fn to_mutable(&self) -> MutDAF<R>

Copies the underlying bytes of this DAF into a MutDAF, enabling modification of the DAF.