anise::naif::daf

Trait NAIFDataSet

pub trait NAIFDataSet<'a>:
    Sized
    + Display
    + PartialEq {
    type RecordKind: NAIFDataRecord<'a>;
    type StateKind;

    const DATASET_NAME: &'static str;

    // Required methods
    fn from_f64_slice(slice: &'a [f64]) -> Result<Self, DecodingError>;
    fn nth_record(&self, n: usize) -> Result<Self::RecordKind, DecodingError>;
    fn evaluate<S: NAIFSummaryRecord>(
        &self,
        epoch: Epoch,
        summary: &S,
    ) -> Result<Self::StateKind, InterpolationError>;
    fn check_integrity(&self) -> Result<(), IntegrityError>;

    // Provided methods
    fn to_f64_daf_vec(&self) -> Result<Vec<f64>, InterpolationError> { ... }
    fn truncate<S: NAIFSummaryRecord>(
        self,
        _summary: &S,
        _new_start: Option<Epoch>,
        _new_end: Option<Epoch>,
    ) -> Result<Self, InterpolationError> { ... }
}

Required Associated Constants

const DATASET_NAME: &'static str

The name of this data set, used in errors

Required Associated Types

type RecordKind: NAIFDataRecord<'a>

The underlying record representation

type StateKind

The state that is returned from an evaluation of this data set

Required Methods

fn from_f64_slice(slice: &'a [f64]) -> Result<Self, DecodingError>

Builds this dataset given a slice of f64 data

fn nth_record(&self, n: usize) -> Result<Self::RecordKind, DecodingError>

fn evaluate<S: NAIFSummaryRecord>( &self, epoch: Epoch, summary: &S, ) -> Result<Self::StateKind, InterpolationError>

fn check_integrity(&self) -> Result<(), IntegrityError>

Checks the integrity of this data set, returns an error if the data has issues.

Provided Methods

fn to_f64_daf_vec(&self) -> Result<Vec<f64>, InterpolationError>

Builds the DAF array representing this data

fn truncate<S: NAIFSummaryRecord>( self, _summary: &S, _new_start: Option<Epoch>, _new_end: Option<Epoch>, ) -> Result<Self, InterpolationError>

Returns a copy of Self where the data corresponds to the start and end times provided. If either is set to None, then that data will not be modified.

Rounding of truncation

The exact new start and new end times may not match the exact data. This function only truncates the available data and does not make any modifications to it (i.e. it does not recompute the interpolation coefficients).

Error

This function will return an error if the new start is before the current start or if the new end is after the current end of the data. That’s because it can’t make up new data.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.