anise::structure::dataset

Struct DataSet

pub struct DataSet<T: DataSetT> {
    pub metadata: Metadata,
    pub lut: LookUpTable,
    pub data_checksum: u32,
    pub data: Vec<T>,
}

A DataSet is the core structure shared by all ANISE binary data.

Fields
metadata: Metadatalut: LookUpTable

All datasets have LookUpTable (LUT) that stores the mapping between a key and its index in the ephemeris list.

data_checksum: u32data: Vec<T>

The actual data from the dataset

Implementations

impl DataSet<PlanetaryData>

pub fn describe(&self) -> String

Returns a table describing this planetary data set

impl DataSet<Location>

pub fn to_dhallset(&self) -> Result<LocationDhallSet, String>

Converts a location dataset kernel into its Dhall representation struct

impl DataSet<EulerParameter>

pub fn describe(&self) -> String

Returns a table describing this planetary data set

impl DataSet<Location>

pub fn describe(&self) -> String

Returns a table describing this planetary data set

impl<T: DataSetT> DataSet<T>

pub fn try_from_bytes<B: Deref<Target = [u8]>>( bytes: B, ) -> Result<Self, DataSetError>

Try to load an Anise file from a pointer of bytes

pub fn from_bytes<B: Deref<Target = [u8]>>(buf: B) -> Self

Forces to load an Anise file from a pointer of bytes. Panics if the bytes cannot be interpreted as an Anise file.

pub fn crc32(&self) -> u32

Compute the CRC32 of the underlying bytes

pub fn set_crc32(&mut self)

Sets the checksum of this data. NOTE: For this calculation, the data checksum field is set to u32::MAX;

pub fn check_integrity(&self) -> Result<(), IntegrityError>

pub fn scrub(&self) -> Result<(), IntegrityError>

Scrubs the data by computing the CRC32 of the bytes and making sure that it still matches the previously known hash

pub fn push( &mut self, item: T, id: Option<i32>, name: Option<&str>, ) -> Result<(), DataSetError>

pub fn get_by_id(&self, id: i32) -> Result<T, DataSetError>

Get a copy of the data with that ID, if that ID is in the lookup table

pub fn set_by_id(&mut self, id: i32, new_value: T) -> Result<(), DataSetError>

Mutates this dataset to change the value of the entry with that ID to the new provided value. This will return an error if the ID is not in the lookup table. Note that this function requires a new heap allocation to change the underlying dataset

pub fn rm_by_id(&mut self, id: i32) -> Result<(), DataSetError>

👎Deprecated since 0.7.0: use clear_by_id instead

pub fn clear_by_id(&mut self, id: i32) -> Result<(), DataSetError>

Mutates this dataset to clear an entry by its ID.

This clears the entry in the data vector by replacing it with its default value, and removes the ID from the look-up table. The corresponding name, if any, is also removed from the look-up table. This will return an error if the ID is not in the lookup table.

pub fn get_by_name(&self, name: &str) -> Result<T, DataSetError>

Get a copy of the data with that name, if that name is in the lookup table

pub fn set_by_name( &mut self, name: &str, new_value: T, ) -> Result<(), DataSetError>

Mutates this dataset to change the value of the entry with that name to the new provided value. This will return an error if the name is not in the lookup table. Note that this function requires a new heap allocation to change the underlying dataset

pub fn rm_by_name(&mut self, name: &str) -> Result<(), DataSetError>

👎Deprecated since 0.7.0: use clear_by_name instead

pub fn clear_by_name(&mut self, name: &str) -> Result<(), DataSetError>

Mutates this dataset to clear an entry by its name.

This clears the entry in the data vector by replacing it with its default value, and removes the name from the look-up table. The corresponding ID, if any, is also removed from the look-up table. This will return an error if the name is not in the lookup table

pub fn save_as( &self, filename: &PathBuf, overwrite: bool, ) -> Result<(), DataSetError>

Saves this dataset to the provided file If overwrite is set to false, and the filename already exists, this function will return an error.

pub fn len(&self) -> usize

Returns the length of the LONGEST of the two look up tables

pub fn is_empty(&self) -> bool

Returns whether this dataset is empty