anise::structure::planetocentric

Struct PlanetaryData

pub struct PlanetaryData {
    pub object_id: i32,
    pub parent_id: i32,
    pub mu_km3_s2: f64,
    pub shape: Option<Ellipsoid>,
    pub pole_right_ascension: Option<PhaseAngle<MAX_NUT_PREC_ANGLES>>,
    pub pole_declination: Option<PhaseAngle<MAX_NUT_PREC_ANGLES>>,
    pub prime_meridian: Option<PhaseAngle<MAX_NUT_PREC_ANGLES>>,
    pub long_axis: Option<f64>,
    pub num_nut_prec_angles: u8,
    pub nut_prec_angles: [PhaseAngle<0>; 32],
}

ANISE supports two different kinds of orientation data. High precision, with spline based interpolations, and constants right ascension, declination, and prime meridian, typically used for planetary constant data.

Documentation of rotation angles

Source: https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/req/pck.html#Models%20for%20the%20Sun,%20Planets,%20and%20some%20Minor%20Bodies%20in%20Text%20PCK%20Kernels The angles RA, DEC, and W are defined as follows:

                                 2
                            RA2*t
RA  =  RA0  +  RA1*t/T  +  ------  + [optional trig polynomials]
                               2
                              T

                                 2
                           DEC2*t
DEC =  DEC0 + DEC1*t/T  +  ------- + [optional trig polynomials]
                               2
                              T

                                2
                            W2*t
W   =  W0   + W1*t/d    +  -----   + [optional trig polynomials]
                              2
                             d

where

d = seconds/day T = seconds/Julian century t = ephemeris time, expressed as seconds past the reference epoch for this body or planetary system

Fields
object_id: i32

The NAIF ID of this object

parent_id: i32

The NAIF ID of the parent orientation, NOT the parent translation

mu_km3_s2: f64

Gravitational parameter (μ) of this planetary object.

shape: Option<Ellipsoid>

The shape is always a tri axial ellipsoid

pole_right_ascension: Option<PhaseAngle<MAX_NUT_PREC_ANGLES>>pole_declination: Option<PhaseAngle<MAX_NUT_PREC_ANGLES>>prime_meridian: Option<PhaseAngle<MAX_NUT_PREC_ANGLES>>long_axis: Option<f64>num_nut_prec_angles: u8

These are the nutation precession angles as a list of tuples to rebuild them. E.g. For E1 = 125.045 - 0.052992 d, this would be stored as a single entry (125.045, -0.052992).

nut_prec_angles: [PhaseAngle<0>; 32]
Implementations

impl PlanetaryData

pub fn to_frame(&self, uid: FrameUid) -> Frame

Converts this planetary data into a Frame, unsetting any shape data for non-body-fixed frames (ID < 100).

pub fn rotation_to_parent( &self, epoch: Epoch, system: &Self, ) -> PhysicsResult<DCM>

Computes the rotation to the parent frame, including its time derivative.

Source: https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/req/rotation.html#Working%20with%20RA,%20Dec%20and%20Twist