anise::structure

Module spacecraft

Structs

DragData

Inertia

Inertial tensor definition

Mass

Defines a spacecraft mass a the sum of the dry (structural) mass and the propellant mass, both in kilogram

SRPData

SpacecraftData

Spacecraft constants can store the some of the spacecraft constant data as the CCSDS Orbit Parameter Message (OPM) and CCSDS Attitude Parameter Messages (APM)