anise

Module time

Re-export of hifitime

Modules

efmt

errors

initializers

leap_seconds

ops

parse

prelude

Structs

Duration

Defines generally usable durations for nanosecond precision valid for 32,768 centuries in either direction, and only on 80 bits / 10 octets.

Epoch

Defines a nanosecond-precision Epoch.

Polynomial

Interpolation Polynomial used for example in [TimeScale] maintenance, precise monitoring or conversions.

TimeSeries

An iterator of a sequence of evenly spaced Epochs.

Enums

DurationError

Freq

An Enum to convert frequencies to their approximate duration, rounded to the closest nanosecond.

HifitimeError

Errors handles all oddities which may occur in this library.

MonthName

Defines Month names, can be initialized either from its variant or its integer (1 for January).

ParsingError

TimeScale

Enum of the different time systems available

Unit

An Enum to perform time unit conversions.

Weekday

Constants

BDT_REF_EPOCH

BDT(BeiDou): 2005 Dec 31st Midnight BDT (BeiDou) reference epoch is 2005 December 31st UTC at midnight. This time scale is synchronized with UTC. |UTC - TAI| = XX Leap Seconds on that day.

DAYS_GPS_TAI_OFFSET

DAYS_PER_CENTURY

DAYS_PER_CENTURY corresponds to the number of days per century in the Julian calendar.

DAYS_PER_CENTURY_I64

DAYS_PER_CENTURY_U64

DAYS_PER_WEEK

DAYS_PER_WEEK_I64

DAYS_PER_YEAR

DAYS_PER_YEAR corresponds to the number of days per year in the Julian calendar.

DAYS_PER_YEAR_NLD

DAYS_PER_YEAR_NLD corresponds to the number of days per year without leap days.

ET_EPOCH_S

The Ephemeris Time epoch, in seconds

GPST_REF_EPOCH

GST_REF_EPOCH

GST (Galileo) reference epoch is 13 seconds before 1999 August 21 UTC at midnight. |UTC - TAI| = XX Leap Seconds on that day.

J1900_REF_EPOCH

The J1900 reference epoch (1900-01-01 at noon) TAI.

J2000_REF_EPOCH

The J2000 reference epoch (2000-01-01 at midnight) TAI. |UTC - TAI| = XX Leap Seconds on that day.

JD_J1900

Julian date for the J1900 epoch, as per NAIF SPICE.

JD_J2000

Julian date for the J2000 epoch, as per NAIF SPICE.

MJD_J1900

Julian days between 01 Jan 1900 at midnight and the Modified Julian Day at 17 November 1858.

MJD_J2000

Julian days between 01 Jan 2000 at noon and the Modified Julian Day at 17 November 1858.

MJD_OFFSET

Modified Julian Date in seconds as defined here. MJD epoch is Modified Julian Day at 17 November 1858 at midnight.

NAIF_EB

NAIF leap second kernel data for EB used to calculate the eccentric anomaly of the heliocentric orbit of the Earth-Moon barycenter.

NAIF_K

NAIF leap second kernel data used to calculate the difference between ET and TAI.

NAIF_M0

NAIF leap second kernel data for M_0 used to calculate the mean anomaly of the heliocentric orbit of the Earth-Moon barycenter.

NAIF_M1

NAIF leap second kernel data for M_1 used to calculate the mean anomaly of the heliocentric orbit of the Earth-Moon barycenter.

NANOSECONDS_PER_CENTURY

NANOSECONDS_PER_DAY

NANOSECONDS_PER_HOUR

NANOSECONDS_PER_MICROSECOND

NANOSECONDS_PER_MILLISECOND

NANOSECONDS_PER_MINUTE

NANOSECONDS_PER_SECOND

QZSST_REF_EPOCH

QZSS and GPS share the same reference epoch.

SECONDS_BDT_TAI_OFFSET

SECONDS_BDT_TAI_OFFSET_I64

SECONDS_GPS_TAI_OFFSET

SECONDS_GPS_TAI_OFFSET_I64

SECONDS_GST_TAI_OFFSET

SECONDS_GST_TAI_OFFSET_I64

SECONDS_PER_CENTURY

SECONDS_PER_CENTURY defines the number of seconds per century.

SECONDS_PER_DAY

SECONDS_PER_DAY defines the number of seconds per day.

SECONDS_PER_DAY_I64

SECONDS_PER_HOUR

SECONDS_PER_HOUR defines the number of seconds per hour.

SECONDS_PER_MINUTE

SECONDS_PER_MINUTE defines the number of seconds per minute.

SECONDS_PER_SIDEREAL_YEAR

SECONDS_PER_SIDEREAL_YEAR corresponds to the number of seconds per sidereal year from NIST.

SECONDS_PER_TROPICAL_YEAR

SECONDS_PER_TROPICAL_YEAR corresponds to the number of seconds per tropical year from NAIF SPICE.

SECONDS_PER_YEAR

SECONDS_PER_YEAR corresponds to the number of seconds per julian year from NAIF SPICE.

SECONDS_PER_YEAR_I64

UNIX_REF_EPOCH

The UNIX reference epoch of 1970-01-01 in TAI duration, accounting only for IERS leap seconds.

Traits

Frequencies

A trait to automatically convert some primitives to an approximate frequency as a duration.

FromStr

Parse a value from a string

TimeUnits

A trait to automatically convert some primitives to a duration

Functions

is_gregorian_valid

Returns true if the provided Gregorian date is valid. Leap second days may have 60 seconds.