anise::time::prelude

Struct Epoch

#[repr(C)]
pub struct Epoch {
    pub duration: Duration,
    pub time_scale: TimeScale,
}

Defines a nanosecond-precision Epoch.

Refer to the appropriate functions for initializing this Epoch from different time scales or representations.

(Python documentation hints) :type string_repr: str

Fields
duration: Duration

An Epoch is always stored as the duration since the beginning of its time scale

time_scale: TimeScale

Time scale used during the initialization of this Epoch.

Implementations

impl Epoch

pub fn to_gregorian_str(&self, time_scale: TimeScale) -> String

Converts the Epoch to Gregorian in the provided time scale and in the ISO8601 format with the time scale appended to the string

pub fn to_gregorian_utc(&self) -> (i32, u8, u8, u8, u8, u8, u32)

Converts the Epoch to the Gregorian UTC equivalent as (year, month, day, hour, minute, second, nanoseconds).

Example

use hifitime::Epoch;

let dt_tai = Epoch::from_tai_parts(1, 537582752000000000);

let dt_str = "2017-01-14T00:31:55 UTC";
let dt = Epoch::from_gregorian_str(dt_str).unwrap();

let (y, m, d, h, min, s, _) = dt_tai.to_gregorian_utc();
assert_eq!(y, 2017);
assert_eq!(m, 1);
assert_eq!(d, 14);
assert_eq!(h, 0);
assert_eq!(min, 31);
assert_eq!(s, 55);
#[cfg(feature = "std")]
{
assert_eq!("2017-01-14T00:31:55 UTC", format!("{dt_tai:?}"));
// dt_tai is initialized from TAI, so the default print is the Gregorian in that time system
assert_eq!("2017-01-14T00:32:32 TAI", format!("{dt_tai}"));
// But dt is initialized from UTC, so the default print and the debug print are both in UTC.
assert_eq!("2017-01-14T00:31:55 UTC", format!("{dt}"));
}

pub fn to_gregorian_tai(&self) -> (i32, u8, u8, u8, u8, u8, u32)

Converts the Epoch to the Gregorian TAI equivalent as (year, month, day, hour, minute, second, nanoseconds).

Example

use hifitime::Epoch;
let dt = Epoch::from_gregorian_tai_at_midnight(1972, 1, 1);
let (y, m, d, h, min, s, _) = dt.to_gregorian_tai();
assert_eq!(y, 1972);
assert_eq!(m, 1);
assert_eq!(d, 1);
assert_eq!(h, 0);
assert_eq!(min, 0);
assert_eq!(s, 0);

pub fn to_gregorian( &self, time_scale: TimeScale, ) -> (i32, u8, u8, u8, u8, u8, u32)

Converts the Epoch to the Gregorian in the provided time scale as (year, month, day, hour, minute, second, nanoseconds).

Example

use hifitime::{Epoch, TimeScale};
let dt = Epoch::from_gregorian_tai_at_midnight(1972, 1, 1);
let (y, m, d, h, min, s, n) = dt.to_gregorian(TimeScale::TAI);
assert_eq!(y, 1972);
assert_eq!(m, 1);
assert_eq!(d, 1);
assert_eq!(h, 0);
assert_eq!(min, 0);
assert_eq!(s, 0);
assert_eq!(n, 0);

// The epoch will be converted to UTC prior to returning the Gregorian parts.
let (y, m, d, h, min, s, n) = dt.to_gregorian(TimeScale::UTC);
assert_eq!(y, 1971);
assert_eq!(m, 12);
assert_eq!(d, 31);
assert_eq!(h, 23);
assert_eq!(min, 59);
assert_eq!(s, 50);
assert_eq!(n, 0);

pub fn maybe_from_gregorian_tai( year: i32, month: u8, day: u8, hour: u8, minute: u8, second: u8, nanos: u32, ) -> Result<Epoch, HifitimeError>

Attempts to build an Epoch from the provided Gregorian date and time in TAI.

pub fn maybe_from_gregorian( year: i32, month: u8, day: u8, hour: u8, minute: u8, second: u8, nanos: u32, time_scale: TimeScale, ) -> Result<Epoch, HifitimeError>

Attempts to build an Epoch from the provided Gregorian date and time in the provided time scale.

Note: The month is ONE indexed, i.e. January is month 1 and December is month 12.

pub fn from_gregorian_tai( year: i32, month: u8, day: u8, hour: u8, minute: u8, second: u8, nanos: u32, ) -> Epoch

Builds an Epoch from the provided Gregorian date and time in TAI. If invalid date is provided, this function will panic. Use maybe_from_gregorian_tai if unsure.

pub fn from_gregorian_tai_at_midnight(year: i32, month: u8, day: u8) -> Epoch

Initialize from the Gregorian date at midnight in TAI.

pub fn from_gregorian_tai_at_noon(year: i32, month: u8, day: u8) -> Epoch

Initialize from the Gregorian date at noon in TAI

pub fn from_gregorian_tai_hms( year: i32, month: u8, day: u8, hour: u8, minute: u8, second: u8, ) -> Epoch

Initialize from the Gregorian date and time (without the nanoseconds) in TAI

pub fn maybe_from_gregorian_utc( year: i32, month: u8, day: u8, hour: u8, minute: u8, second: u8, nanos: u32, ) -> Result<Epoch, HifitimeError>

Attempts to build an Epoch from the provided Gregorian date and time in UTC.

pub fn from_gregorian_utc( year: i32, month: u8, day: u8, hour: u8, minute: u8, second: u8, nanos: u32, ) -> Epoch

Builds an Epoch from the provided Gregorian date and time in UTC. If invalid date is provided, this function will panic. Use maybe_from_gregorian_utc if unsure.

pub fn from_gregorian_utc_at_midnight(year: i32, month: u8, day: u8) -> Epoch

Initialize from Gregorian date in UTC at midnight

pub fn from_gregorian_utc_at_noon(year: i32, month: u8, day: u8) -> Epoch

Initialize from Gregorian date in UTC at noon

pub fn from_gregorian_utc_hms( year: i32, month: u8, day: u8, hour: u8, minute: u8, second: u8, ) -> Epoch

Initialize from the Gregorian date and time (without the nanoseconds) in UTC

pub fn from_gregorian( year: i32, month: u8, day: u8, hour: u8, minute: u8, second: u8, nanos: u32, time_scale: TimeScale, ) -> Epoch

Builds an Epoch from the provided Gregorian date and time in the provided time scale. If invalid date is provided, this function will panic. Use maybe_from_gregorian if unsure.

pub fn from_gregorian_at_midnight( year: i32, month: u8, day: u8, time_scale: TimeScale, ) -> Epoch

Initialize from Gregorian date in UTC at midnight

pub fn from_gregorian_at_noon( year: i32, month: u8, day: u8, time_scale: TimeScale, ) -> Epoch

Initialize from Gregorian date in UTC at noon

pub fn from_gregorian_hms( year: i32, month: u8, day: u8, hour: u8, minute: u8, second: u8, time_scale: TimeScale, ) -> Epoch

Initialize from the Gregorian date and time (without the nanoseconds) in UTC

pub fn from_gregorian_str(s_in: &str) -> Result<Epoch, HifitimeError>

Converts a Gregorian date time in ISO8601 or RFC3339 format into an Epoch, accounting for the time zone designator and the time scale.

Definition

Time Zone Designator: this is either a Z (lower or upper case) to specify UTC, or an offset in hours and minutes off of UTC, such as +01:00 for UTC plus one hour and zero minutes.

Time system (or time “scale”): UTC, TT, TAI, TDB, ET, etc.

Converts an ISO8601 or RFC3339 datetime representation to an Epoch. If no time scale is specified, then UTC is assumed. A time scale may be specified in addition to the format unless The T which separates the date from the time can be replaced with a single whitespace character (\W). The offset is also optional, cf. the examples below.

Example

use hifitime::Epoch;
let dt = Epoch::from_gregorian_utc(2017, 1, 14, 0, 31, 55, 0);
assert_eq!(
    dt,
    Epoch::from_gregorian_str("2017-01-14T00:31:55 UTC").unwrap()
);
assert_eq!(
    dt,
    Epoch::from_gregorian_str("2017-01-14T00:31:55.0000 UTC").unwrap()
);
assert_eq!(
    dt,
    Epoch::from_gregorian_str("2017-01-14T00:31:55").unwrap()
);
assert_eq!(
    dt,
    Epoch::from_gregorian_str("2017-01-14 00:31:55").unwrap()
);
// Regression test for #90
assert_eq!(
    Epoch::from_gregorian_utc(2017, 1, 14, 0, 31, 55, 811000000),
    Epoch::from_gregorian_str("2017-01-14 00:31:55.811 UTC").unwrap()
);
assert_eq!(
    Epoch::from_gregorian_utc(2017, 1, 14, 0, 31, 55, 811200000),
    Epoch::from_gregorian_str("2017-01-14 00:31:55.8112 UTC").unwrap()
);
// Example from https://www.w3.org/TR/NOTE-datetime
assert_eq!(
    Epoch::from_gregorian_utc_hms(1994, 11, 5, 13, 15, 30),
    Epoch::from_gregorian_str("1994-11-05T13:15:30Z").unwrap()
);
assert_eq!(
    Epoch::from_gregorian_utc_hms(1994, 11, 5, 13, 15, 30),
    Epoch::from_gregorian_str("1994-11-05T08:15:30-05:00").unwrap()
);

impl Epoch

pub const fn from_tai_duration(duration: Duration) -> Epoch

Creates a new Epoch from a Duration as the time difference between this epoch and TAI reference epoch.

pub fn to_duration_since_j1900(&self) -> Duration

pub fn from_tai_parts(centuries: i16, nanoseconds: u64) -> Epoch

Creates a new Epoch from its centuries and nanosecond since the TAI reference epoch.

pub fn from_tai_seconds(seconds: f64) -> Epoch

Initialize an Epoch from the provided TAI seconds since 1900 January 01 at midnight

pub fn from_tai_days(days: f64) -> Epoch

Initialize an Epoch from the provided TAI days since 1900 January 01 at midnight

pub fn from_utc_duration(duration: Duration) -> Epoch

Initialize an Epoch from the provided UTC seconds since 1900 January 01 at midnight

pub fn from_utc_seconds(seconds: f64) -> Epoch

Initialize an Epoch from the provided UTC seconds since 1900 January 01 at midnight

pub fn from_utc_days(days: f64) -> Epoch

Initialize an Epoch from the provided UTC days since 1900 January 01 at midnight

pub fn from_gpst_duration(duration: Duration) -> Epoch

Initialize an Epoch from the provided duration since 1980 January 6 at midnight

pub fn from_qzsst_duration(duration: Duration) -> Epoch

Initialize an Epoch from the provided duration since 1980 January 6 at midnight

pub fn from_gst_duration(duration: Duration) -> Epoch

Initialize an Epoch from the provided duration since August 21st 1999 midnight

pub fn from_bdt_duration(duration: Duration) -> Epoch

Initialize an Epoch from the provided duration since January 1st midnight

pub fn from_mjd_tai(days: f64) -> Epoch

pub fn from_mjd_in_time_scale(days: f64, time_scale: TimeScale) -> Epoch

pub fn from_mjd_utc(days: f64) -> Epoch

pub fn from_mjd_gpst(days: f64) -> Epoch

pub fn from_mjd_qzsst(days: f64) -> Epoch

pub fn from_mjd_gst(days: f64) -> Epoch

pub fn from_mjd_bdt(days: f64) -> Epoch

pub fn from_jde_tai(days: f64) -> Epoch

pub fn from_jde_in_time_scale(days: f64, time_scale: TimeScale) -> Epoch

pub fn from_jde_utc(days: f64) -> Epoch

pub fn from_jde_gpst(days: f64) -> Epoch

pub fn from_jde_qzsst(days: f64) -> Epoch

pub fn from_jde_gst(days: f64) -> Epoch

pub fn from_jde_bdt(days: f64) -> Epoch

pub fn from_tt_seconds(seconds: f64) -> Epoch

Initialize an Epoch from the provided TT seconds (approximated to 32.184s delta from TAI)

pub fn from_tt_duration(duration: Duration) -> Epoch

Initialize an Epoch from the provided TT seconds (approximated to 32.184s delta from TAI)

pub fn from_et_seconds(seconds_since_j2000: f64) -> Epoch

Initialize an Epoch from the Ephemeris Time seconds past 2000 JAN 01 (J2000 reference)

pub fn from_et_duration(duration_since_j2000: Duration) -> Epoch

Initializes an Epoch from the duration between J2000 and the current epoch as per NAIF SPICE.

Limitation

This method uses a Newton Raphson iteration to find the appropriate TAI duration. This method is only accuracy to a few nanoseconds. Hence, when calling as_et_duration() and re-initializing it with from_et_duration you may have a few nanoseconds of difference (expect less than 10 ns).

Warning

The et2utc function of NAIF SPICE will assume that there are 9 leap seconds before 01 JAN 1972, as this date introduces 10 leap seconds. At the time of writing, this does not seem to be in line with IERS and the documentation in the leap seconds list.

In order to match SPICE, the as_et_duration() function will manually get rid of that difference.

pub fn from_tdb_seconds(seconds_j2000: f64) -> Epoch

Initialize an Epoch from Dynamic Barycentric Time (TDB) seconds past 2000 JAN 01 midnight (difference than SPICE) NOTE: This uses the ESA algorithm, which is a notch more complicated than the SPICE algorithm, but more precise. In fact, SPICE algorithm is precise +/- 30 microseconds for a century whereas ESA algorithm should be exactly correct.

pub fn from_tdb_duration(duration_since_j2000: Duration) -> Epoch

Initialize from Dynamic Barycentric Time (TDB) (same as SPICE ephemeris time) whose epoch is 2000 JAN 01 noon TAI.

pub fn from_jde_et(days: f64) -> Epoch

Initialize from the JDE days

pub fn from_jde_tdb(days: f64) -> Epoch

Initialize from Dynamic Barycentric Time (TDB) (same as SPICE ephemeris time) in JD days

pub fn from_gpst_seconds(seconds: f64) -> Epoch

Initialize an Epoch from the number of seconds since the GPS Time Epoch, defined as UTC midnight of January 5th to 6th 1980 (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS#GPS_Time_.28GPST.29).

pub fn from_gpst_days(days: f64) -> Epoch

Initialize an Epoch from the number of days since the GPS Time Epoch, defined as UTC midnight of January 5th to 6th 1980 (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS#GPS_Time_.28GPST.29).

pub fn from_gpst_nanoseconds(nanoseconds: u64) -> Epoch

Initialize an Epoch from the number of nanoseconds since the GPS Time Epoch, defined as UTC midnight of January 5th to 6th 1980 (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS#GPS_Time_.28GPST.29). This may be useful for time keeping devices that use GPS as a time source.

pub fn from_qzsst_seconds(seconds: f64) -> Epoch

Initialize an Epoch from the number of seconds since the QZSS Time Epoch, defined as UTC midnight of January 5th to 6th 1980 (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS#GPS_Time_.28GPST.29).

pub fn from_qzsst_days(days: f64) -> Epoch

Initialize an Epoch from the number of days since the QZSS Time Epoch, defined as UTC midnight of January 5th to 6th 1980 (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS#GPS_Time_.28GPST.29).

pub fn from_qzsst_nanoseconds(nanoseconds: u64) -> Epoch

Initialize an Epoch from the number of nanoseconds since the QZSS Time Epoch, defined as UTC midnight of January 5th to 6th 1980 (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS#GPS_Time_.28GPST.29). This may be useful for time keeping devices that use QZSS as a time source.

pub fn from_gst_seconds(seconds: f64) -> Epoch

Initialize an Epoch from the number of seconds since the GST Time Epoch, starting August 21st 1999 midnight (UTC) (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS).

pub fn from_gst_days(days: f64) -> Epoch

Initialize an Epoch from the number of days since the GST Time Epoch, starting August 21st 1999 midnight (UTC) (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS)

pub fn from_gst_nanoseconds(nanoseconds: u64) -> Epoch

Initialize an Epoch from the number of nanoseconds since the GPS Time Epoch, starting August 21st 1999 midnight (UTC) (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS)

pub fn from_bdt_seconds(seconds: f64) -> Epoch

Initialize an Epoch from the number of seconds since the BDT Time Epoch, starting on January 1st 2006 (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS)

pub fn from_bdt_days(days: f64) -> Epoch

Initialize an Epoch from the number of days since the BDT Time Epoch, starting on January 1st 2006 (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS)

pub fn from_bdt_nanoseconds(nanoseconds: u64) -> Epoch

Initialize an Epoch from the number of nanoseconds since the BDT Time Epoch, starting on January 1st 2006 (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS). This may be useful for time keeping devices that use BDT as a time source.

pub fn from_ptp_duration(duration: Duration) -> Epoch

Initialize an Epoch from the provided IEEE 1588-2008 (PTPv2) duration since TAI midnight 1970 January 01. PTP uses the TAI timescale but with the Unix Epoch for compatibility with unix systems.

pub fn from_ptp_seconds(seconds: f64) -> Epoch

Initialize an Epoch from the provided IEEE 1588-2008 (PTPv2) second timestamp since TAI midnight 1970 January 01. PTP uses the TAI timescale but with the Unix Epoch for compatibility with unix systems.

pub fn from_ptp_nanoseconds(nanoseconds: u64) -> Epoch

Initialize an Epoch from the provided IEEE 1588-2008 (PTPv2) nanoseconds timestamp since TAI midnight 1970 January 01. PTP uses the TAI timescale but with the Unix Epoch for compatibility with unix systems.

pub fn from_unix_duration(duration: Duration) -> Epoch

Initialize an Epoch from the provided duration since UTC midnight 1970 January 01.

pub fn from_unix_seconds(seconds: f64) -> Epoch

Initialize an Epoch from the provided UNIX second timestamp since UTC midnight 1970 January 01.

pub fn from_unix_milliseconds(millisecond: f64) -> Epoch

Initialize an Epoch from the provided UNIX millisecond timestamp since UTC midnight 1970 January 01.

pub fn from_str_with_format( s_in: &str, format: Format, ) -> Result<Epoch, HifitimeError>

Initializes an Epoch from the provided Format.

pub fn from_format_str( s_in: &str, format_str: &str, ) -> Result<Epoch, HifitimeError>

Initializes an Epoch from the Format as a string.

pub fn from_time_of_week( week: u32, nanoseconds: u64, time_scale: TimeScale, ) -> Epoch

Builds an Epoch from given week: elapsed weeks counter into the desired Time scale, and the amount of nanoseconds within that week. For example, this is how GPS vehicles describe a GPST epoch.

Note that this constructor relies on 128 bit integer math and may be slow on embedded devices.

pub fn from_time_of_week_utc(week: u32, nanoseconds: u64) -> Epoch

Builds a UTC Epoch from given week: elapsed weeks counter and “ns” amount of nanoseconds since closest Sunday Midnight.

pub fn from_day_of_year(year: i32, days: f64, time_scale: TimeScale) -> Epoch

Builds an Epoch from the provided year, days in the year, and a time scale.

Limitations

In the TDB or ET time scales, there may be an error of up to 750 nanoseconds when initializing an Epoch this way. This is because we first initialize the epoch in Gregorian scale and then apply the TDB/ET offset, but that offset actually depends on the precise time.

Day couting behavior

The day counter starts at 01, in other words, 01 January is day 1 of the counter, as per the GPS specificiations.

impl Epoch

pub fn min(&self, other: Epoch) -> Epoch

Returns the minimum of the two epochs.

use hifitime::Epoch;

let e0 = Epoch::from_gregorian_utc_at_midnight(2022, 10, 20);
let e1 = Epoch::from_gregorian_utc_at_midnight(2022, 10, 21);

assert_eq!(e0, e1.min(e0));
assert_eq!(e0, e0.min(e1));

Note: this uses a pointer to self which will be copied immediately because Python requires a pointer. :type other: Epoch :rtype: Epoch

pub fn max(&self, other: Epoch) -> Epoch

Returns the maximum of the two epochs.

use hifitime::Epoch;

let e0 = Epoch::from_gregorian_utc_at_midnight(2022, 10, 20);
let e1 = Epoch::from_gregorian_utc_at_midnight(2022, 10, 21);

assert_eq!(e1, e1.max(e0));
assert_eq!(e1, e0.max(e1));

Note: this uses a pointer to self which will be copied immediately because Python requires a pointer. :type other: Epoch :rtype: Epoch

pub fn floor(&self, duration: Duration) -> Epoch

Floors this epoch to the closest provided duration

Example

use hifitime::{Epoch, TimeUnits};

let e = Epoch::from_gregorian_tai_hms(2022, 5, 20, 17, 57, 43);
assert_eq!(
    e.floor(1.hours()),
    Epoch::from_gregorian_tai_hms(2022, 5, 20, 17, 0, 0)
);

let e = Epoch::from_gregorian_tai(2022, 10, 3, 17, 44, 29, 898032665);
assert_eq!(
    e.floor(3.minutes()),
    Epoch::from_gregorian_tai_hms(2022, 10, 3, 17, 42, 0)
);

:type duration: Duration :rtype: Epoch

pub fn ceil(&self, duration: Duration) -> Epoch

Ceils this epoch to the closest provided duration in the TAI time scale

Example

use hifitime::{Epoch, TimeUnits};

let e = Epoch::from_gregorian_tai_hms(2022, 5, 20, 17, 57, 43);
assert_eq!(
    e.ceil(1.hours()),
    Epoch::from_gregorian_tai_hms(2022, 5, 20, 18, 0, 0)
);

// 45 minutes is a multiple of 3 minutes, hence this result
let e = Epoch::from_gregorian_tai(2022, 10, 3, 17, 44, 29, 898032665);
assert_eq!(
    e.ceil(3.minutes()),
    Epoch::from_gregorian_tai_hms(2022, 10, 3, 17, 45, 0)
);

:type duration: Duration :rtype: Epoch

pub fn round(&self, duration: Duration) -> Epoch

Rounds this epoch to the closest provided duration in TAI

Example

use hifitime::{Epoch, TimeUnits};

let e = Epoch::from_gregorian_tai_hms(2022, 5, 20, 17, 57, 43);
assert_eq!(
    e.round(1.hours()),
    Epoch::from_gregorian_tai_hms(2022, 5, 20, 18, 0, 0)
);

:type duration: Duration :rtype: Epoch

pub fn to_time_of_week(&self) -> (u32, u64)

Converts this epoch into the time of week, represented as a rolling week counter into that time scale and the number of nanoseconds elapsed in current week (since closest Sunday midnight). This is usually how GNSS receivers describe a timestamp. :rtype: tuple

pub fn precise_timescale_conversion( &self, forward: bool, reference_epoch: Epoch, polynomial: Polynomial, target: TimeScale, ) -> Result<Epoch, HifitimeError>

Converts this Epoch into targeted TimeScale using provided Polynomial.

Input

forward: whether this is forward or backward conversion. For example, using GPST-UTC Polynomial

GPST->UTC is the forward conversion

UTC->GPST is the backward conversion

reference_epoch: any reference Epoch for the provided Polynomial.

While we support any time difference, it should remain short in pratice (a day at most, for precise applications).

polynomial: that must be valid for this reference Epoch, used in the equation a0 + a1*dt + a2*dt² = GPST-UTC for example.

target: targetted TimeScale we will transition to.

Example:

use hifitime::{Epoch, TimeScale, Polynomial, Unit};

// random GPST Epoch for forward conversion to UTC
let t_gpst = Epoch::from_gregorian(2020, 01, 01, 0, 0, 0, 0, TimeScale::GPST);

// Let's say we know the GPST-UTC polynomials for that day,
// They allow precise forward transition from GPST to UTC,
// and precise backward transition from UTC to GPST.
let gpst_utc_polynomials = Polynomial::from_constant_offset_nanoseconds(1.0);

// This is the reference [Epoch] attached to the publication of these polynomials.
// You should use polynomials that remain valid and were provided recently (usually one day at most).
// Example: polynomials were published 1 hour ago.
let gpst_reference = t_gpst - 1.0 * Unit::Hour;

// Forward conversion (to UTC) GPST - a0 + a1 *dt + a2*dt² = UTC
let t_utc = t_gpst.precise_timescale_conversion(true, gpst_reference, gpst_utc_polynomials, TimeScale::UTC)
    .unwrap();

// Verify we did transition to UTC
assert_eq!(t_utc.time_scale, TimeScale::UTC);

// Verify the resulting [Epoch] is the coarse GPST->UTC transition + fine correction
let reversed = t_utc.to_time_scale(TimeScale::GPST) + 1.0 * Unit::Nanosecond;
assert_eq!(reversed, t_gpst);

// Apply the backward transition, from t_utc back to t_gpst.
// The timescale conversion works both ways: (from UTC) GPST = UTC + a0 + a1 *dt + a2*dt²
let backwards = t_utc.precise_timescale_conversion(false, gpst_reference, gpst_utc_polynomials, TimeScale::GPST)
    .unwrap();

assert_eq!(backwards, t_gpst);

// It is important to understand that your reference point does not have to be in the past.
// The only logic that should prevail is to always minimize interpolation gap.
// In other words, if you can access future interpolation information that would minimize the data gap, they should prevail.
// Example: +30' in the future.
let gpst_reference = t_gpst + 30.0 * Unit::Minute;

// Forward conversion (to UTC) but using polynomials that were released 1 hour after t_gpst
let t_utc = t_gpst.precise_timescale_conversion(true, gpst_reference, gpst_utc_polynomials, TimeScale::UTC)
    .unwrap();

// Verifications
assert_eq!(t_utc.time_scale, TimeScale::UTC);

let reversed = t_utc.to_time_scale(TimeScale::GPST) + 1.0 * Unit::Nanosecond;
assert_eq!(reversed, t_gpst);

:type forward: bool :type reference_epoch: Epoch :type polynomial: Polynomial :type target: TimeScale :rtype: Epoch

pub fn weekday_in_time_scale(&self, time_scale: TimeScale) -> Weekday

Returns the weekday in provided time scale ASSUMING that the reference epoch of that time scale is a Monday. You probably do not want to use this. You probably either want weekday() or weekday_utc(). Several time scales do not have a reference day that’s on a Monday, e.g. BDT. :type time_scale: TimeScale :rtype: Weekday

pub fn weekday(&self) -> Weekday

Returns weekday (uses the TAI representation for this calculation). :rtype: Weekday

pub fn weekday_utc(&self) -> Weekday

Returns weekday in UTC timescale :rtype: Weekday

pub fn next(&self, weekday: Weekday) -> Epoch

Returns the next weekday.

use hifitime::prelude::*;

let epoch = Epoch::from_gregorian_utc_at_midnight(1988, 1, 2);
assert_eq!(epoch.weekday_utc(), Weekday::Saturday);
assert_eq!(epoch.next(Weekday::Sunday), Epoch::from_gregorian_utc_at_midnight(1988, 1, 3));
assert_eq!(epoch.next(Weekday::Monday), Epoch::from_gregorian_utc_at_midnight(1988, 1, 4));
assert_eq!(epoch.next(Weekday::Tuesday), Epoch::from_gregorian_utc_at_midnight(1988, 1, 5));
assert_eq!(epoch.next(Weekday::Wednesday), Epoch::from_gregorian_utc_at_midnight(1988, 1, 6));
assert_eq!(epoch.next(Weekday::Thursday), Epoch::from_gregorian_utc_at_midnight(1988, 1, 7));
assert_eq!(epoch.next(Weekday::Friday), Epoch::from_gregorian_utc_at_midnight(1988, 1, 8));
assert_eq!(epoch.next(Weekday::Saturday), Epoch::from_gregorian_utc_at_midnight(1988, 1, 9));

:type weekday: Weekday :rtype: Epoch

pub fn next_weekday_at_midnight(&self, weekday: Weekday) -> Epoch

:type weekday: Weekday :rtype: Epoch

pub fn next_weekday_at_noon(&self, weekday: Weekday) -> Epoch

:type weekday: Weekday :rtype: Epoch

pub fn previous(&self, weekday: Weekday) -> Epoch

Returns the next weekday.

use hifitime::prelude::*;

let epoch = Epoch::from_gregorian_utc_at_midnight(1988, 1, 2);
assert_eq!(epoch.previous(Weekday::Friday), Epoch::from_gregorian_utc_at_midnight(1988, 1, 1));
assert_eq!(epoch.previous(Weekday::Thursday), Epoch::from_gregorian_utc_at_midnight(1987, 12, 31));
assert_eq!(epoch.previous(Weekday::Wednesday), Epoch::from_gregorian_utc_at_midnight(1987, 12, 30));
assert_eq!(epoch.previous(Weekday::Tuesday), Epoch::from_gregorian_utc_at_midnight(1987, 12, 29));
assert_eq!(epoch.previous(Weekday::Monday), Epoch::from_gregorian_utc_at_midnight(1987, 12, 28));
assert_eq!(epoch.previous(Weekday::Sunday), Epoch::from_gregorian_utc_at_midnight(1987, 12, 27));
assert_eq!(epoch.previous(Weekday::Saturday), Epoch::from_gregorian_utc_at_midnight(1987, 12, 26));

:type weekday: Weekday :rtype: Epoch

pub fn previous_weekday_at_midnight(&self, weekday: Weekday) -> Epoch

:type weekday: Weekday :rtype: Epoch

pub fn previous_weekday_at_noon(&self, weekday: Weekday) -> Epoch

:type weekday: Weekday :rtype: Epoch

impl Epoch

pub fn with_hms(&self, hours: u64, minutes: u64, seconds: u64) -> Epoch

Returns a copy of self where the time is set to the provided hours, minutes, seconds Invalid number of hours, minutes, and seconds will overflow into their higher unit. Warning: this does not set the subdivisions of second to zero. :type hours: int :type minutes: int :type seconds: int :rtype: Epoch

pub fn with_hms_from(&self, other: Epoch) -> Epoch

Returns a copy of self where the hours, minutes, seconds is set to the time of the provided epoch but the sub-second parts are kept from the current epoch.

:type other: Epoch :rtype: Epoch

use hifitime::prelude::*;

let epoch = Epoch::from_gregorian_utc(2022, 12, 01, 10, 11, 12, 13);
let other_utc = Epoch::from_gregorian_utc(2024, 12, 01, 20, 21, 22, 23);
let other = other_utc.to_time_scale(TimeScale::TDB);

assert_eq!(
    epoch.with_hms_from(other),
    Epoch::from_gregorian_utc(2022, 12, 01, 20, 21, 22, 13)
);

pub fn with_time_from(&self, other: Epoch) -> Epoch

Returns a copy of self where all of the time components (hours, minutes, seconds, and sub-seconds) are set to the time of the provided epoch.

:type other: Epoch :rtype: Epoch

use hifitime::prelude::*;

let epoch = Epoch::from_gregorian_utc(2022, 12, 01, 10, 11, 12, 13);
let other_utc = Epoch::from_gregorian_utc(2024, 12, 01, 20, 21, 22, 23);
// If the other Epoch is in another time scale, it does not matter, it will be converted to the correct time scale.
let other = other_utc.to_time_scale(TimeScale::TDB);

assert_eq!(
    epoch.with_time_from(other),
    Epoch::from_gregorian_utc(2022, 12, 01, 20, 21, 22, 23)
);

pub fn with_hms_strict(&self, hours: u64, minutes: u64, seconds: u64) -> Epoch

Returns a copy of self where the time is set to the provided hours, minutes, seconds Invalid number of hours, minutes, and seconds will overflow into their higher unit. Warning: this will set the subdivisions of seconds to zero. :type hours: int :type minutes: int :type seconds: int :rtype: Epoch

pub fn with_hms_strict_from(&self, other: Epoch) -> Epoch

Returns a copy of self where the time is set to the time of the other epoch but the subseconds are set to zero.

:type other: Epoch :rtype: Epoch

use hifitime::prelude::*;

let epoch = Epoch::from_gregorian_utc(2022, 12, 01, 10, 11, 12, 13);
let other_utc = Epoch::from_gregorian_utc(2024, 12, 01, 20, 21, 22, 23);
let other = other_utc.to_time_scale(TimeScale::TDB);

assert_eq!(
    epoch.with_hms_strict_from(other),
    Epoch::from_gregorian_utc(2022, 12, 01, 20, 21, 22, 0)
);

impl Epoch

pub fn now() -> Result<Epoch, HifitimeError>

Initializes a new Epoch from now. WARNING: This assumes that the system time returns the time in UTC (which is the case on Linux) Uses std::time::SystemTime::now or javascript interop under the hood

impl Epoch

pub fn leap_seconds_with<L>(&self, iers_only: bool, provider: L) -> Option<f64>
where L: LeapSecondProvider,

Get the accumulated number of leap seconds up to this Epoch from the provided LeapSecondProvider. Returns None if the epoch is before 1960, year at which UTC was defined.

Why does this function return an Option when the other returns a value

This is to match the iauDat function of SOFA (src/dat.c). That function will return a warning and give up if the start date is before 1960.

pub const fn from_duration(duration: Duration, ts: TimeScale) -> Epoch

Creates an epoch from given duration expressed in given timescale, i.e. since the given time scale’s reference epoch.

For example, if the duration is 1 day and the time scale is Ephemeris Time, then this will create an epoch of 2000-01-02 at midnight ET. If the duration is 1 day and the time scale is TAI, this will create an epoch of 1900-01-02 at noon, because the TAI reference epoch in Hifitime is chosen to be the J1900 epoch. In case of ET, TDB Timescales, a duration since J2000 is expected.

impl Epoch

pub fn to_time_scale(&self, ts: TimeScale) -> Epoch

Converts self to another time scale

As per the Rust naming convention, this borrows an Epoch and returns an owned Epoch.

:type ts: TimeScale :rtype: Epoch

pub fn leap_seconds_iers(&self) -> i32

Get the accumulated number of leap seconds up to this Epoch accounting only for the IERS leap seconds. :rtype: int

pub fn leap_seconds(&self, iers_only: bool) -> Option<f64>

Get the accumulated number of leap seconds up to this Epoch accounting only for the IERS leap seconds and the SOFA scaling from 1960 to 1972, depending on flag. Returns None if the epoch is before 1960, year at which UTC was defined.

Why does this function return an Option when the other returns a value

This is to match the iauDat function of SOFA (src/dat.c). That function will return a warning and give up if the start date is before 1960. :type iers_only: bool :rtype: float

pub fn to_isoformat(&self) -> String

The standard ISO format of this epoch (six digits of subseconds) in the current time scale, refer to https://docs.rs/hifitime/latest/hifitime/efmt/format/struct.Format.html for format options. :rtype: str

pub fn to_duration_in_time_scale(&self, ts: TimeScale) -> Duration

Returns this epoch with respect to the provided time scale. This is needed to correctly perform duration conversions in dynamical time scales (e.g. TDB). :type ts: TimeScale :rtype: Duration

pub fn to_tai_seconds(&self) -> f64

Returns the number of TAI seconds since J1900 :rtype: float

pub fn to_tai_duration(&self) -> Duration

Returns this time in a Duration past J1900 counted in TAI :rtype: Duration

pub fn to_tai(&self, unit: Unit) -> f64

Returns the epoch as a floating point value in the provided unit :type unit: Unit :rtype: float

pub fn to_tai_parts(&self) -> (i16, u64)

Returns the TAI parts of this duration :rtype: tuple

pub fn to_tai_days(&self) -> f64

Returns the number of days since J1900 in TAI :rtype: float

pub fn to_utc_seconds(&self) -> f64

Returns the number of UTC seconds since the TAI epoch :rtype: float

pub fn to_utc_duration(&self) -> Duration

Returns this time in a Duration past J1900 counted in UTC :rtype: Duration

pub fn to_utc(&self, unit: Unit) -> f64

Returns the number of UTC seconds since the TAI epoch :type unit: Unit :rtype: float

pub fn to_utc_days(&self) -> f64

Returns the number of UTC days since the TAI epoch :rtype: float

pub fn to_mjd_tai_days(&self) -> f64

as_mjd_days creates an Epoch from the provided Modified Julian Date in days as explained here. MJD epoch is Modified Julian Day at 17 November 1858 at midnight. :rtype: float

pub fn to_mjd_tai_seconds(&self) -> f64

Returns the Modified Julian Date in seconds TAI. :rtype: float

pub fn to_mjd_tai(&self, unit: Unit) -> f64

Returns this epoch as a duration in the requested units in MJD TAI :type unit: Unit :rtype: float

pub fn to_mjd_utc_days(&self) -> f64

Returns the Modified Julian Date in days UTC. :rtype: float

pub fn to_mjd_utc(&self, unit: Unit) -> f64

Returns the Modified Julian Date in the provided unit in UTC. :type unit: Unit :rtype: float

pub fn to_mjd_utc_seconds(&self) -> f64

Returns the Modified Julian Date in seconds UTC. :rtype: float

pub fn to_jde_tai_days(&self) -> f64

Returns the Julian days from epoch 01 Jan -4713, 12:00 (noon) as explained in “Fundamentals of astrodynamics and applications”, Vallado et al. 4th edition, page 182, and on Wikipedia. :rtype: float

pub fn to_jde_tai(&self, unit: Unit) -> f64

Returns the Julian Days from epoch 01 Jan -4713 12:00 (noon) in desired Duration::Unit :type unit: Unit :rtype: float

pub fn to_jde_tai_duration(&self) -> Duration

Returns the Julian Days from epoch 01 Jan -4713 12:00 (noon) as a Duration :rtype: Duration

pub fn to_jde_tai_seconds(&self) -> f64

Returns the Julian seconds in TAI. :rtype: float

pub fn to_jde_utc_days(&self) -> f64

Returns the Julian days in UTC. :rtype: float

pub fn to_jde_utc_duration(&self) -> Duration

Returns the Julian days in UTC as a Duration :rtype: Duration

pub fn to_jde_utc_seconds(&self) -> f64

Returns the Julian Days in UTC seconds. :rtype: float

pub fn to_tt_seconds(&self) -> f64

Returns seconds past TAI epoch in Terrestrial Time (TT) (previously called Terrestrial Dynamical Time (TDT)) :rtype: float

pub fn to_tt_duration(&self) -> Duration

Returns Duration past TAI epoch in Terrestrial Time (TT). :rtype: Duration

pub fn to_tt_days(&self) -> f64

Returns days past TAI epoch in Terrestrial Time (TT) (previously called Terrestrial Dynamical Time (TDT)) :rtype: float

pub fn to_tt_centuries_j2k(&self) -> f64

Returns the centuries passed J2000 TT :rtype: float

pub fn to_tt_since_j2k(&self) -> Duration

Returns the duration past J2000 TT :rtype: Duration

pub fn to_jde_tt_days(&self) -> f64

Returns days past Julian epoch in Terrestrial Time (TT) (previously called Terrestrial Dynamical Time (TDT)) :rtype: float

pub fn to_jde_tt_duration(&self) -> Duration

:rtype: Duration

pub fn to_mjd_tt_days(&self) -> f64

Returns days past Modified Julian epoch in Terrestrial Time (TT) (previously called Terrestrial Dynamical Time (TDT)) :rtype: float

pub fn to_mjd_tt_duration(&self) -> Duration

:rtype: Duration

pub fn to_gpst_seconds(&self) -> f64

Returns seconds past GPS Time Epoch, defined as UTC midnight of January 5th to 6th 1980 (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS#GPS_Time_.28GPST.29). :rtype: float

pub fn to_gpst_duration(&self) -> Duration

Returns Duration past GPS time Epoch. :rtype: Duration

pub fn to_gpst_nanoseconds(&self) -> Result<u64, HifitimeError>

Returns nanoseconds past GPS Time Epoch, defined as UTC midnight of January 5th to 6th 1980 (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS#GPS_Time_.28GPST.29). NOTE: This function will return an error if the centuries past GPST time are not zero. :rtype: int

pub fn to_gpst_days(&self) -> f64

Returns days past GPS Time Epoch, defined as UTC midnight of January 5th to 6th 1980 (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS#GPS_Time_.28GPST.29). :rtype: float

pub fn to_qzsst_seconds(&self) -> f64

Returns seconds past QZSS Time Epoch, defined as UTC midnight of January 5th to 6th 1980 (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS#GPS_Time_.28GPST.29). :rtype: float

pub fn to_qzsst_duration(&self) -> Duration

Returns Duration past QZSS time Epoch. :rtype: Duration

pub fn to_qzsst_nanoseconds(&self) -> Result<u64, HifitimeError>

Returns nanoseconds past QZSS Time Epoch, defined as UTC midnight of January 5th to 6th 1980 (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS#GPS_Time_.28GPST.29). NOTE: This function will return an error if the centuries past QZSST time are not zero. :rtype: int

pub fn to_qzsst_days(&self) -> f64

Returns days past QZSS Time Epoch, defined as UTC midnight of January 5th to 6th 1980 (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS#GPS_Time_.28GPST.29). :rtype: float

pub fn to_gst_seconds(&self) -> f64

Returns seconds past GST (Galileo) Time Epoch :rtype: float

pub fn to_gst_duration(&self) -> Duration

Returns Duration past GST (Galileo) time Epoch. :rtype: Duration

pub fn to_gst_nanoseconds(&self) -> Result<u64, HifitimeError>

Returns nanoseconds past GST (Galileo) Time Epoch, starting on August 21st 1999 Midnight UT (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS). NOTE: This function will return an error if the centuries past GST time are not zero. :rtype: int

pub fn to_gst_days(&self) -> f64

Returns days past GST (Galileo) Time Epoch, starting on August 21st 1999 Midnight UT (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS). :rtype: float

pub fn to_bdt_seconds(&self) -> f64

Returns seconds past BDT (BeiDou) Time Epoch :rtype: float

pub fn to_bdt_duration(&self) -> Duration

Returns Duration past BDT (BeiDou) time Epoch. :rtype: Duration

pub fn to_bdt_days(&self) -> f64

Returns days past BDT (BeiDou) Time Epoch, defined as Jan 01 2006 UTC (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS). :rtype: float

pub fn to_bdt_nanoseconds(&self) -> Result<u64, HifitimeError>

Returns nanoseconds past BDT (BeiDou) Time Epoch, defined as Jan 01 2006 UTC (cf. https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS). NOTE: This function will return an error if the centuries past GST time are not zero. :rtype: int

pub fn to_unix(&self, unit: Unit) -> f64

Returns the duration since the UNIX epoch in the provided unit. :type unit: Unit :rtype: float

pub fn to_unix_seconds(&self) -> f64

Returns the number seconds since the UNIX epoch defined 01 Jan 1970 midnight UTC. :rtype: float

pub fn to_unix_milliseconds(&self) -> f64

Returns the number milliseconds since the UNIX epoch defined 01 Jan 1970 midnight UTC. :rtype: float

pub fn to_unix_days(&self) -> f64

Returns the number days since the UNIX epoch defined 01 Jan 1970 midnight UTC. :rtype: float

pub fn to_et_seconds(&self) -> f64

Returns the Ephemeris Time seconds past 2000 JAN 01 midnight, matches NASA/NAIF SPICE. :rtype: float

pub fn to_et_duration(&self) -> Duration

Returns the duration between J2000 and the current epoch as per NAIF SPICE.

Warning

The et2utc function of NAIF SPICE will assume that there are 9 leap seconds before 01 JAN 1972, as this date introduces 10 leap seconds. At the time of writing, this does not seem to be in line with IERS and the documentation in the leap seconds list.

In order to match SPICE, the as_et_duration() function will manually get rid of that difference. :rtype: Duration

pub fn to_tdb_duration(&self) -> Duration

Returns the Dynamics Barycentric Time (TDB) as a high precision Duration since J2000

Algorithm

Given the embedded sine functions in the equation to compute the difference between TDB and TAI from the number of TDB seconds past J2000, one cannot solve the revert the operation analytically. Instead, we iterate until the value no longer changes.

Assume that the TAI duration is in fact the TDB seconds from J2000.

Offset to J2000 because Epoch stores everything in the J1900 but the TDB duration is in J2000.

Compute the offset g due to the TDB computation with the current value of the TDB seconds (defined in step 1).

Subtract that offset to the latest TDB seconds and store this as a new candidate for the true TDB seconds value.

Compute the difference between this candidate and the previous one. If the difference is less than one nanosecond, stop iteration.

Set the new candidate as the TDB seconds since J2000 and loop until step 5 breaks the loop, or we’ve done five iterations.

At this stage, we have a good approximation of the TDB seconds since J2000.

Reverse the algorithm given that approximation: compute the g offset, compute the difference between TDB and TAI, add the TT offset (32.184 s), and offset by the difference between J1900 and J2000.

:rtype: Duration

pub fn to_tdb_seconds(&self) -> f64

Returns the Dynamic Barycentric Time (TDB) (higher fidelity SPICE ephemeris time) whose epoch is 2000 JAN 01 noon TAI (cf. https://gssc.esa.int/navipedia/index.php/Transformations_between_Time_Systems#TDT_-_TDB.2C_TCB) :rtype: float

pub fn to_jde_et_days(&self) -> f64

Returns the Ephemeris Time JDE past epoch :rtype: float

pub fn to_jde_et_duration(&self) -> Duration

:rtype: Duration

pub fn to_jde_et(&self, unit: Unit) -> f64

:type unit: Unit :rtype: float

pub fn to_jde_tdb_duration(&self) -> Duration

:rtype: Duration

pub fn to_jde_tdb_days(&self) -> f64

Returns the Dynamic Barycentric Time (TDB) (higher fidelity SPICE ephemeris time) whose epoch is 2000 JAN 01 noon TAI (cf. https://gssc.esa.int/navipedia/index.php/Transformations_between_Time_Systems#TDT_-_TDB.2C_TCB) :rtype: float

pub fn to_tdb_days_since_j2000(&self) -> f64

Returns the number of days since Dynamic Barycentric Time (TDB) J2000 (used for Archinal et al. rotations) :rtype: float

pub fn to_tdb_centuries_since_j2000(&self) -> f64

Returns the number of centuries since Dynamic Barycentric Time (TDB) J2000 (used for Archinal et al. rotations) :rtype: float

pub fn to_et_days_since_j2000(&self) -> f64

Returns the number of days since Ephemeris Time (ET) J2000 (used for Archinal et al. rotations) :rtype: float

pub fn to_et_centuries_since_j2000(&self) -> f64

Returns the number of centuries since Ephemeris Time (ET) J2000 (used for Archinal et al. rotations) :rtype: float

pub fn duration_in_year(&self) -> Duration

Returns the duration since the start of the year :rtype: Duration

pub fn day_of_year(&self) -> f64

Returns the number of days since the start of the year. :rtype: float

pub fn day_of_month(&self) -> u8

Returns the number of days since the start of the Gregorian month in the current time scale.

Example

use hifitime::Epoch;
let dt = Epoch::from_gregorian_tai_at_midnight(2025, 7, 3);
assert_eq!(dt.day_of_month(), 3);

:rtype: int

pub fn year(&self) -> i32

Returns the number of Gregorian years of this epoch in the current time scale. :rtype: int

pub fn year_days_of_year(&self) -> (i32, f64)

Returns the year and the days in the year so far (days of year). :rtype: tuple

pub fn hours(&self) -> u64

Returns the hours of the Gregorian representation of this epoch in the time scale it was initialized in. :rtype: int

pub fn minutes(&self) -> u64

Returns the minutes of the Gregorian representation of this epoch in the time scale it was initialized in. :rtype: int

pub fn seconds(&self) -> u64

Returns the seconds of the Gregorian representation of this epoch in the time scale it was initialized in. :rtype: int

pub fn milliseconds(&self) -> u64

Returns the milliseconds of the Gregorian representation of this epoch in the time scale it was initialized in. :rtype: int

pub fn microseconds(&self) -> u64

Returns the microseconds of the Gregorian representation of this epoch in the time scale it was initialized in. :rtype: int

pub fn nanoseconds(&self) -> u64

Returns the nanoseconds of the Gregorian representation of this epoch in the time scale it was initialized in. :rtype: int

pub fn month_name(&self) -> MonthName

:rtype: MonthName

pub fn to_rfc3339(&self) -> String

Returns this epoch in UTC in the RFC3339 format :rtype: str